==> 5. Baseline Local Models.ipynb <==
# Clear workspace
rm(list=ls())

# Garabage collect to help with memory issues
gc()

# Uncomment and run to install packages if needed
# install.packages("tidyverse")
# install.packages("cluster") 
# install.packages("forecast") 
# install.packages("Metrics")
# install.packages("tseries")
# install.packages("scoringutils")

library(tidyverse)
library(parallel)
library(cluster)
library(forecast)
library(Metrics)
library(tseries)
library(scoringutils)

# Create the folders to save results in
res_folder1 <- "Results/Univariate/Auto Arima/"
if (!file.exists(res_folder1)) {
 dir.create(res_folder1, recursive=TRUE)
}

res_folder2 <- "Results/Univariate/Naive/"
if (!file.exists(res_folder2)) {
 dir.create(res_folder2, recursive=TRUE)
}


# Read In and Subset Data

# Read in all files from the Data/Intermediate directory 
# and create a list of data frames for each highway system
england_fnames <- list.files("Data/Processed/Highways_England/", pattern="*.csv", full.names=TRUE)
england_df_list <- lapply(1:length(england_fnames), 
                          function(x) read_csv(england_fnames[x]) %>% arrange(timestamp))

portland_fnames <- list.files("Data/Processed/Portland/", pattern="*.csv", full.names=TRUE)
portland_df_list <- lapply(1:length(portland_fnames), 
                          function(x) read_csv(portland_fnames[x]) %>% arrange(timestamp))

utah_fnames <- list.files("Data/Processed/Utah/", pattern="*.csv", full.names=TRUE)
utah_df_list <- lapply(1:length(utah_fnames), 
                          function(x) read_csv(utah_fnames[x]) %>% arrange(timestamp))

# Create one total list of all data frames
total_df_list <- append(append(england_df_list, portland_df_list), utah_df_list)

# Ensure that the rows in each data frame are in proper chronological order - this should have been taken care of
# in a previous notebook, but it cannot hurt to make sure things are correct here as well
england_df_list <- lapply(england_df_list, function(x) x %>% arrange(timestamp))
portland_df_list <- lapply(portland_df_list, function(x) x %>% arrange(timestamp))
utah_df_list <- lapply(utah_df_list, function(x) x %>% arrange(timestamp))    

# Create one list of all data frames
total_df_list <- append(append(england_df_list, portland_df_list), utah_df_list)

# Grab the csv of starting and ending points for sampling each time series
start_end_points <- read_csv('start_end_points.csv')
starting_points <- start_end_points$start
ending_points <- start_end_points$end

# For each df in our list, create a row number column called 'rn' - this will allow us to find the starting point
# for each sample using the above starting_points array
total_df_list <- lapply(total_df_list, function(x) x %>% mutate(rn = row_number()))

# Sample each df in the list according to its corresponding starting and ending point - this creates 12-week
# long samples of each data frame which will be used for modeling and testing
total_df_list_samples <- lapply(1:length(total_df_list), 
                                function(x) total_df_list[[x]] %>% 
                                    filter(rn >= starting_points[x]) %>%
                                    filter(rn <= ending_points[x]))

# Add a new column to each data frame to denote whether each row should be part of the training set, validation set
# or test set - the first 8 weeks are designated as train, the next 2 as val, and the final 2 as test. It may be 
# the case that train and val get used for training, depending on the modeling task
total_df_list_samples <- lapply(total_df_list_samples, 
                                function(x) x %>% mutate(rn = row_number()) %>%
                                    mutate(train_val_test = ifelse(rn <= (96*7*8), 
                                                                   "train", 
                                                                   ifelse(rn <= (96*7*10), 
                                                                          "val", 
                                                                          "test")))
                                
                               )

# Create a list of data frames which only have the training and validation rows of each df
train_val_samples <- lapply(total_df_list_samples, function(x) x %>% filter(train_val_test != "test"))

# Create a list of data frames which only contain the test samples for each df
test_samples <- lapply(total_df_list_samples, function(x) x %>% filter(train_val_test == "test"))

# Create a list of time series objects from the total_volume column of each data frame using the msts function
# This function allows for time series with multiple seasonality, so we use both daily and weekly seasonal periods
# Having time series objects instead of data frame columns is helpful for some tasks
# Here, we create this list for the train_val samples list, next for the test samples
train_val_samples_ts <- lapply(train_val_samples, 
                               function(x) msts(x$total_volume, seasonal.periods=c(24*4, 24*4*7)))

# Create a similar list for the test data
test_samples_ts <- lapply(test_samples,
                          function(x) msts(x$total_volume, seasonal.periods=c(24*4, 24*4*7)))

# ARIMA with Fourier Terms

arima_w_fourier <- function(ts_train) { 
    # Function which takes in training time series object, creates Fourier terms for that ts,  
    # and returns a model using auto.arima and fourier terms as regressors
    
    # Create Fourier terms with K=2 for both frequencies/seasonal periods
    fourier_term_train <- fourier(ts_train, K=c(2,2))
    
    # Use auto.arima with default parameters and seasonal set to FALSE 
    # as the Fourier terms account for the seasonality
    arima_model <- auto.arima(ts_train, xreg = fourier_term_train, seasonal = FALSE) 
    
    # Return the model from auto.arim
    arima_model
}

# Call the above arima_w_fourier function for each train_val sample in our data set
# Because we are allowing auto.arime to choose the final modeling structure for each time series, we
# are passing the full train and validation time series together. We could also manually compute the best
# model structure by creating models on the training time frames and computing forecast accuracy on the validation
# time frames. It would be an interesting avenue of future work to test whether this makes any difference 
# in terms of model performance on the final test sets
arima_res <- lapply(1:length(train_val_samples_ts), 
                    function(x) arima_w_fourier(train_val_samples_ts[[x]]))

head(arima_res)

arima_res[[36]]

# Save the ARIMA results to an R data file so they can be used later without having to re-run models
saveRDS(arima_res, file="Results/Univariate/Auto Arima/auto_arima_models.RData")

# Read in results from the above modeling
arima_res <- readRDS("Results/Univariate/Auto Arima/auto_arima_models.RData")

arima_w_fourier_predict <- function(ts_train, ts_test, model) {
    # Function to make one-step ahead forecasts given a model, the data the model was trained on, and testing data
    
    # Create the fourier terms for training and testing sets
    fourier_term_train <- fourier(ts_train, K=c(2,2))
    fourier_term_test <- fourier(ts_train, K=c(2,2), h=length(ts_test))

    # Make one-step ahead forecasts on the training data using the fitted function
    train_preds <- fitted(model)
    
    # Compute the rmse and mae of the one-step ahead predictions on the training data
    train_rmse <- rmse(ts_train, train_preds)
    train_mae <- mae(ts_train, train_preds)
    
    # Create a "test" version of the model by adding in the test data and calling Arima with the full data set
    # and the already constructed model
    mod_test <- Arima(y=c(ts_train, ts_test), 
                      xreg=rbind(fourier_term_train,
                                 fourier_term_test),
                      model=model)
    
    # Compute test preds by calling fitted on the new model and eliminating the training preds component of the 
    # results
    test_preds <- fitted(mod_test)[-c(1:length(ts_train))]
    
    # Compute the rmse and mae on the one-step ahead forecasts on the test data
    test_rmse <- rmse(ts_test, test_preds)
    test_mae <- mae(ts_test, test_preds)
    
    # Return a list of the rmse and mae for train and test sets, as well as the actual forecasts on the test data
    return_list <- list("train_rmse"=train_rmse, "train_mae"=train_mae,
                        "test_rmse"=test_rmse, "test_mae"=test_mae, "raw_forecasts"=test_preds)
}

# Call the above function for each time series in our data set and its corresponding ARIMA model
arima_perf <- lapply(1:length(test_samples_ts),
                     function(x) arima_w_fourier_predict(train_val_samples_ts[[x]],
                                                         test_samples_ts[[x]],
                                                         arima_res[[x]])
                    )

# Create a data frame from the above results, with one column for each train and test rmse and mae
arima_perf_df <- data.frame(train_rmse=unlist(lapply(arima_perf, function(x) x$train_rmse)),
                            train_mae=unlist(lapply(arima_perf, function(x) x$train_mae)),
                            test_rmse=unlist(lapply(arima_perf, function(x) x$test_rmse)),
                            test_mae=unlist(lapply(arima_perf, function(x) x$test_mae)))

# Save the results to a CSV and an R data file
write.csv(arima_perf_df, "Results/Univariate/Auto Arima/auto_arima_performance.csv", row.names=FALSE)
saveRDS(arima_perf_df, file="Results/Univariate/Auto Arima/auto_arima_performance.RData")

# Read in the results from the R data file
arima_results <- readRDS("Results/Univariate/Auto Arima/auto_arima_performance.RData")

# In the next several cells, print and inspect the performance of each model on each time series, including
# examining rmse, mae, and the normalized/standardized versions of each by first computing the mean of the time
# series and scaling appropriately
mean(arima_results$train_rmse)

mean(arima_results$train_mae)

mean(arima_results$test_rmse)

mean(arima_results$test_mae)

# Compute means for scaling RMSE and MAE
test_means <- unlist(lapply(test_samples_ts, mean))
train_means <- unlist(lapply(train_val_samples_ts, mean))

mean(arima_results$train_rmse/train_means)

mean(arima_results$train_mae/train_means)

mean(arima_results$test_rmse/test_means)

mean(arima_results$test_mae/test_means)

# Create a list of just the one-step ahead forecasts for each test set
arima_raw_forecasts <- lapply(1:length(arima_perf), 
                              function(x) data.frame(forecast=arima_perf[[x]]$raw_forecasts, ts_index=x))

# Take that list and turn it into a data frame
arima_raw_forecasts_df <- do.call("rbind", arima_raw_forecasts)

# Print head of the df
head(arima_raw_forecasts_df)

# Check df length - just a sanity check to make sure things look alright
nrow(arima_raw_forecasts_df)

# Save the raw forecasts to a CSV file for future use
write.csv(arima_raw_forecasts_df, "Results/Univariate/Auto Arima/arima_raw_forecasts.csv", row.names=FALSE)

# Read in the raw forecast results
arima_raw_forecasts_df <- read_csv("Results/Univariate/Auto Arima/arima_raw_forecasts.csv")

bootstrap_pred_int <- function(predictions,
                               residuals,
                               n_boot,
                               actuals=NULL
                              ) 
{
    # Function to compute prediction intervals using bootstrapping with model residuals
    # function takes in the predicted values, the model residuals, the number of bootstrap samples to use,
    # and an optional set of true values (used for computing prediction interval scores)
    
    # Intialize empty lists for the hi and lo components of the 80% nad 95% PIs
    boot_lo_80 <- c()
    boot_hi_80 <- c()
    boot_lo_95 <- c()
    boot_hi_95 <- c()

    # Loop through every row in the predictions
    for (i in 1:nrow(predictions)){
        # Reset the seed to the loop counter - could make this more random, but this is for reproducibility
        set.seed(i)
        
        # Sample with replacement from the residuals n_boot times
        boot_samples <- sample(residuals, n_boot, replace=TRUE)
        # Add the predicted value to each sample
        boot_preds <- boot_samples + predictions$forecast[i]
        
        # Compute the PIs as the quantiles of these pred+boot values
        boot_pred_lo_80 <- quantile(boot_preds, 0.1)
        boot_pred_hi_80 <- quantile(boot_preds, 0.9)

        boot_pred_lo_95 <- quantile(boot_preds, 0.025)
        boot_pred_hi_95 <- quantile(boot_preds, 0.975)
        
        # Append the PI values to the lists
        boot_lo_80 <- c(boot_lo_80, boot_pred_lo_80)
        boot_hi_80 <- c(boot_hi_80, boot_pred_hi_80)

        boot_lo_95 <- c(boot_lo_95, boot_pred_lo_95)
        boot_hi_95 <- c(boot_hi_95, boot_pred_hi_95)        
    }
    
    # Create a copy of the input df called predicitons
    pred_w_boot_int_df <- predictions
    
    # Add columns to the df for the PIs
    pred_w_boot_int_df$lo_80 <- boot_lo_80
    pred_w_boot_int_df$hi_80 <- boot_hi_80
    pred_w_boot_int_df$lo_95 <- boot_lo_95
    pred_w_boot_int_df$hi_95 <- boot_hi_95
    
    # If the true values are included
    if (!is.null(actuals)){
        
        # Add a column for the true values
        pred_w_boot_int_df$true_values <- actuals
        
        # Compute the 80% and 95% interval scores and add them as columns in the df
        pred_w_boot_int_df <- pred_w_boot_int_df %>%
            mutate(int_score_80 = interval_score(true_values=true_values,
                                                 lower=lo_80,
                                                 upper=hi_80,
                                                 interval_range=80,
                                                 weigh=FALSE
                                                )) %>% 
            mutate(int_score_95 = interval_score(true_values=true_values,
                                                 lower=lo_95,
                                                 upper=hi_95,
                                                 interval_range=95,
                                                 weigh=FALSE
                                                ))
    }
    
    # Return the final set of preds with PIs (and interval scores if applicable)
    pred_w_boot_int_df
    
}

# Set the number of bootstrap samples to use
n_boot <- 1000

# Compute bootstrap PI for each model/ts
bootstrap_pi <- lapply(1:length(arima_raw_forecasts), 
                      function(i) bootstrap_pred_int(arima_raw_forecasts[[i]], 
                                                     arima_res[[i]]$residuals, 
                                                     n_boot, 
                                                     test_samples_ts[[i]]))

# Print the mean of the 80% PI scores
mean(unlist(lapply(bootstrap_pi, function(x) mean(x$int_score_80))))

# Print the mean of the 95% PI scores
mean(unlist(lapply(bootstrap_pi, function(x) mean(x$int_score_95))))

mean(unlist(lapply(bootstrap_pi, function(x) mean(x$int_score_80)))/test_means)

mean(unlist(lapply(bootstrap_pi, function(x) mean(x$int_score_95)))/test_means)

# Save the bootstrap PIs to an RDS file
saveRDS(bootstrap_pi, file="Results/Univariate/Auto Arima/auto_arima_bootstrap_pi.RData")

# Naive Forecasts

naive_forec <- function(train_ts, test_ts, lag=1) {
    # Function to make "naive" forecasts using a given lag, and then compute the accuracy metrics and residuals
    
    # Train forecasts are simply computed as the value at the time series lag steps ago
    train_forec <- lag(train_ts, lag)
    
    # Compute the residuals on the train set, removing the first lag values because they will result in nulls
    train_res <- train_ts[-c(1:lag)] - train_forec[-c(1:lag)]
        
    # Copmute the rmse and mae on the train set
    train_rmse <- rmse(train_ts[-c(1:lag)], train_forec[-c(1:lag)])
    train_mae <- mae(train_ts[-c(1:lag)], train_forec[-c(1:lag)])
    
    # Append the final time steps of the train and test data together
    n <- length(train_ts)
    n1 <- n-(lag-1)
    train_test_data <- c(train_ts[n1:n], test_ts)
    
    # Compute test forecasts from this appended data set
    test_forec <- lag(train_test_data, lag)
    test_forec_final <- test_forec[-c(1:lag)]
        
    # Compute rmse and mae on the test data
    test_rmse <- rmse(test_ts, test_forec_final)
    test_mae <- mae(test_ts, test_forec_final)
        
    # Return a list which includes the rmse, mae for both train and test, the residuals on the train data,
    # and the raw forecasts on the test data
    return_list <- list("train_rmse"=train_rmse, "train_mae"=train_mae, "train_residuals"=train_res,
                        "test_rmse"=test_rmse, "test_mae"=test_mae, "raw_forecasts"=test_forec_final) 

}

bootstrap_pred_int_naive <- function(predictions,
                                     residuals,
                                     n_boot,
                                     actuals=NULL
                                    ) 
{
    # Function to compute prediction intervals using bootstrapping with model residuals
    # function takes in the predicted values, the model residuals, the number of bootstrap samples to use,
    # and an optional set of true values (used for computing prediction interval scores)
    
    # Intialize empty lists for the hi and lo components of the 80% nad 95% PIs
    boot_lo_80 <- c()
    boot_hi_80 <- c()
    boot_lo_95 <- c()
    boot_hi_95 <- c()

    # Loop through every row in the predictions
    for (i in 1:length(predictions)){
        # Reset the seed to the loop counter - could make this more random, but this is for reproducibility
        set.seed(i)
        
        # Sample with replacement from the residuals n_boot times
        boot_samples <- sample(residuals, n_boot, replace=TRUE)
        # Add the predicted value to each sample
        boot_preds <- boot_samples + predictions[i]
        
        # Compute the PIs as the quantiles of these pred+boot values
        boot_pred_lo_80 <- quantile(boot_preds, 0.1)
        boot_pred_hi_80 <- quantile(boot_preds, 0.9)

        boot_pred_lo_95 <- quantile(boot_preds, 0.025)
        boot_pred_hi_95 <- quantile(boot_preds, 0.975)
        
        # Append the PI values to the lists
        boot_lo_80 <- c(boot_lo_80, boot_pred_lo_80)
        boot_hi_80 <- c(boot_hi_80, boot_pred_hi_80)

        boot_lo_95 <- c(boot_lo_95, boot_pred_lo_95)
        boot_hi_95 <- c(boot_hi_95, boot_pred_hi_95)        
    }
    
    # Create a copy of the input df called predicitons
    pred_w_boot_int_df <- data.frame(forecast=predictions)
    
    # Add columns to the df for the PIs
    pred_w_boot_int_df$lo_80 <- boot_lo_80
    pred_w_boot_int_df$hi_80 <- boot_hi_80
    pred_w_boot_int_df$lo_95 <- boot_lo_95
    pred_w_boot_int_df$hi_95 <- boot_hi_95
    
    # If the true values are included
    if (!is.null(actuals)){
        
        # Add a column for the true values
        pred_w_boot_int_df$true_values <- actuals
        
        # Compute the 80% and 95% interval scores and add them as columns in the df
        pred_w_boot_int_df <- pred_w_boot_int_df %>%
            mutate(int_score_80 = interval_score(true_values=true_values,
                                                 lower=lo_80,
                                                 upper=hi_80,
                                                 interval_range=80,
                                                 weigh=FALSE
                                                )) %>% 
            mutate(int_score_95 = interval_score(true_values=true_values,
                                                 lower=lo_95,
                                                 upper=hi_95,
                                                 interval_range=95,
                                                 weigh=FALSE
                                                ))
    }
    
    # Return the final set of preds with PIs (and interval scores if applicable)
    pred_w_boot_int_df
    
}

## Simple Naive Forecast

# Apply the above function to compute simple naive forecasts - that is, lag 1 naive forecasts
simple_naive_forecasts <- lapply(1:length(train_val_samples),
                                 function(x) naive_forec(train_val_samples[[x]]$total_volume,
                                                         test_samples[[x]]$total_volume))

# Compute average performance across the data set
mean(unlist(lapply(simple_naive_forecasts, function(x) x$test_rmse)))

mean(unlist(lapply(simple_naive_forecasts, function(x) x$test_mae)))

mean(unlist(lapply(simple_naive_forecasts, function(x) x$test_rmse))/test_means)

mean(unlist(lapply(simple_naive_forecasts, function(x) x$test_mae))/test_means)

# Compute bootstrap PI for each model/ts
bootstrap_pi_naive <- lapply(1:length(simple_naive_forecasts), 
                      function(i) bootstrap_pred_int_naive(simple_naive_forecasts[[i]]$raw_forecasts,
                                                           simple_naive_forecasts[[i]]$train_res, 
                                                           n_boot, 
                                                           test_samples_ts[[i]]))

# Print the mean of the 80% PI scores
mean(unlist(lapply(bootstrap_pi_naive, function(x) mean(x$int_score_80))))

# Print the mean of the 95% PI scores
mean(unlist(lapply(bootstrap_pi_naive, function(x) mean(x$int_score_95))))

mean(unlist(lapply(bootstrap_pi_naive, function(x) mean(x$int_score_80)))/test_means)

mean(unlist(lapply(bootstrap_pi_naive, function(x) mean(x$int_score_95)))/test_means)

# Save the bootstrap PIs to an RDS file
saveRDS(bootstrap_pi_naive, file="Results/Univariate/Naive/naive_bootstrap_pi.RData")

## Daily Seasonal Naive Forecast

# Use the same approach, but this time with lag=96 to use a daily naive forecast instead of a simple one
daily_naive_forecasts <- lapply(1:length(train_val_samples),
                                function(x) naive_forec(train_val_samples[[x]]$total_volume,
                                                        test_samples[[x]]$total_volume,
                                                        lag=96
                                                       )
                                )

mean(unlist(lapply(daily_naive_forecasts, function(x) x$test_rmse)))

mean(unlist(lapply(daily_naive_forecasts, function(x) x$test_mae)))

mean(unlist(lapply(daily_naive_forecasts, function(x) x$test_rmse))/test_means)

mean(unlist(lapply(daily_naive_forecasts, function(x) x$test_mae))/test_means)

# Compute bootstrap PI for each model/ts
bootstrap_pi_naive_daily <- lapply(1:length(daily_naive_forecasts), 
                      function(i) bootstrap_pred_int_naive(daily_naive_forecasts[[i]]$raw_forecasts,
                                                           daily_naive_forecasts[[i]]$train_res, 
                                                           n_boot, 
                                                           test_samples_ts[[i]]))

# Print the mean of the 80% PI scores
mean(unlist(lapply(bootstrap_pi_naive_daily, function(x) mean(x$int_score_80))))

# Print the mean of the 95% PI scores
mean(unlist(lapply(bootstrap_pi_naive_daily, function(x) mean(x$int_score_95))))

mean(unlist(lapply(bootstrap_pi_naive_daily, function(x) mean(x$int_score_80)))/test_means)

mean(unlist(lapply(bootstrap_pi_naive_daily, function(x) mean(x$int_score_95)))/test_means)

# Save the bootstrap PIs to an RDS file
saveRDS(bootstrap_pi_naive, file="Results/Univariate/Naive/daily_naive_bootstrap_pi.RData")

## Weekly Naive Forecast

# Again, use the same methods but this time weekly so lag=672
weekly_naive_forecasts <- lapply(1:length(train_val_samples),
                                 function(x) naive_forec(train_val_samples[[x]]$total_volume,
                                                         test_samples[[x]]$total_volume,
                                                         lag=672
                                                       )
                                )

mean(unlist(lapply(weekly_naive_forecasts, function(x) x$test_rmse)))

mean(unlist(lapply(weekly_naive_forecasts, function(x) x$test_mae)))

mean(unlist(lapply(weekly_naive_forecasts, function(x) x$test_rmse))/test_means)

mean(unlist(lapply(weekly_naive_forecasts, function(x) x$test_mae))/test_means)

# Compute bootstrap PI for each model/ts
bootstrap_pi_naive_weekly <- lapply(1:length(weekly_naive_forecasts), 
                      function(i) bootstrap_pred_int_naive(weekly_naive_forecasts[[i]]$raw_forecasts,
                                                           weekly_naive_forecasts[[i]]$train_res, 
                                                           n_boot, 
                                                           test_samples_ts[[i]]))

# Print the mean of the 80% PI scores
mean(unlist(lapply(bootstrap_pi_naive_weekly, function(x) mean(x$int_score_80))))

# Print the mean of the 95% PI scores
mean(unlist(lapply(bootstrap_pi_naive_weekly, function(x) mean(x$int_score_95))))

mean(unlist(lapply(bootstrap_pi_naive_weekly, function(x) mean(x$int_score_80)))/test_means)

mean(unlist(lapply(bootstrap_pi_naive_weekly, function(x) mean(x$int_score_95)))/test_means)

# Save the bootstrap PIs to an RDS file
saveRDS(bootstrap_pi_naive, file="Results/Univariate/Naive/weekly_naive_bootstrap_pi.RData")
