==> 13. Unseen Sensor - Assign New Data to a Cluster.ipynb <==
# Empty the work space
rm(list=ls())

# Uncomment and run to install packages if needed
# install.packages("tidyverse")
# install.packages("cluster")
# install.packages("tsfeatures")
# install.packages("Rcatch22")
# install.packages("tseries")
# install.packages("factoextra")
# install.packages("forecast")
# install.packages("dtwclust")

# Load libraries
library(tidyverse)
library(cluster)
library(tsfeatures)
library(Rcatch22)
library(tseries)
library(factoextra)
library(forecast)
library(dtwclust)

# Create the folders to save results in
res_folder1 <- "Results/Unseen Sensor/"
if (!file.exists(res_folder1)) {
 dir.create(res_folder1, recursive=TRUE)
}

# Read In Data

# Read in each file into a list of data frames
fnames <- list.files("Data/Unseen Sensor/Processed/", pattern="*.csv", full.names=TRUE)
total_df_list <- lapply(fnames, read_csv)

# Sanity check the lengths of each list
length(total_df_list)

# Ensure all data is arranged in time stamp order
total_df_list <- lapply(total_df_list, function(x) x %>% arrange(timestamp))    

# Randomly create a list of starting points from which we we sample each data frame

# Set the seed
set.seed(12345)

# Sample without replacement as to ensure the starting time for each time series is different
starting_points <- sample(1:((96*365)-(96*12*7)), 
                          length(total_df_list), 
                          replace=FALSE)

# Create the ending points by adding on 12 weeks to the starting points
ending_points <- starting_points + (96*7*12) - 1

# Save the sampling start and end points to a csv for use in other notebooks
write.csv(data.frame(start=starting_points, end=ending_points), "start_end_points_unseen.csv", row.names=F)

# For each df in our list, create a row number column called 'rn' - this will allow us to find the starting point
# for each sample using the above starting_points array
total_df_list <- lapply(total_df_list, function(x) x %>% mutate(rn = row_number()))

# Sample each df in the list according to its corresponding starting and ending point - this creates 12-week
# long samples of each data frame which will be used for modeling and testing
total_df_list_samples <- lapply(1:length(total_df_list), 
                                function(x) total_df_list[[x]] %>% 
                                    filter(rn >= starting_points[x]) %>%
                                    filter(rn <= ending_points[x]))

# Add a new column to each data frame to denote whether each row should be part of the training set, validation set
# or test set - the first 8 weeks are designated as train, the next 2 as val, and the final 2 as test. It may be 
# the case that train and val get used for training, depending on the modeling task
total_df_list_samples <- lapply(total_df_list_samples, 
                                function(x) x %>% mutate(rn = row_number()) %>%
                                    mutate(train_val_test = ifelse(rn <= (96*7*8), 
                                                                   "train", 
                                                                   ifelse(rn <= (96*7*10), 
                                                                          "val", 
                                                                          "test")))
                                
                               )

# Create a list of data frames which only have the training and validation rows of each df
train_val_samples <- lapply(total_df_list_samples, function(x) x %>% filter(train_val_test != "test"))

# Create a list of data frames which only have the training rows of each df
train_samples <- lapply(total_df_list_samples, function(x) x %>% filter(train_val_test == "train"))

# Create a list of data frames which only have the test rows of each df
test_samples <- lapply(total_df_list_samples, function(x) x %>% filter(train_val_test == "test"))

# Get the target field (total_volume) from the dfs in the above lists and create msts objects instead of dfs
train_val_samples_ts <- lapply(train_val_samples, 
                               function(x) msts(x$total_volume, seasonal.periods=c(24*4, 24*4*7)))

# Get the target field (total_volume) from the dfs in the above lists and create msts objects instead of dfs
train_samples_ts <- lapply(train_samples, 
                           function(x) msts(x$total_volume, seasonal.periods=c(24*4, 24*4*7)))

# Get the target field (total_volume) from the dfs in the above lists and create msts objects instead of dfs
test_samples_ts <- lapply(test_samples,
                          function(x) msts(x$total_volume, seasonal.periods=c(24*4, 24*4*7)))

# Random Cluster Assignments

# Total number of time series in the data set
total_clust <- 4

# Assign each time series to a random cluster

# Create a list to save the assignments
rand_clust_assign <- c()
# Loop through a list 1:n, where n is the number of training samples
for (ts_no in 1:length(train_samples_ts)){
    # Set a seed
    set.seed(ts_no)
    # Sample from 1:n and append to the list
    rand_clust_assign <- c(rand_clust_assign, sample(1:total_clust, 1))
}

# Print the list of assignments
rand_clust_assign

# Catch22 Based Feature Computation

# For each train time series in our list, compute the Catch22 features using the catch22_all function 
# and save the results in a df
train_catch22_feat <- lapply(lapply(train_samples_ts, function(x) catch22_all(x)), 
                             function(x) x %>% 
                                 pivot_wider(names_from=names, values_from=values))

# Append all features for each time series into a data frame and remove the features which we did not use 
# previously (in notebook 7) due to low variance
train_catch22_feat_df <- do.call("rbind", train_catch22_feat) %>% 
    select(-c('CO_FirstMin_ac', 'PD_PeriodicityWang_th0_01'))
train_catch22_feat_df

## Read in Catch22 Feature Values for Original Data

# Read in the original catch22 features on the original training data, 
# and remove the first column as it is just an indexer
catch22_orig <- read_csv("Results/Clustering/KMeans/catch22_feat_df.csv")[,-1]

### Scale New Data with Original Scaler Values

# Create a new data frame to save scaled features on the new data
train_catch22_feat_df_scaled <- train_catch22_feat_df

# Loop through each feature
for (n in 1:ncol(train_catch22_feat_df_scaled)){
    # Compute the min of the feature on the original data
    min_n <- min(catch22_orig[,n])
    # Compute the max of the feature on the original data
    max_n <- max(catch22_orig[,n])
    # Scale the feature of the new data using the min and max of the original data
    train_catch22_feat_df_scaled[,n] <- (train_catch22_feat_df_scaled[,n] - min_n)/(max_n - min_n)
}

# Add row names
row.names(train_catch22_feat_df_scaled) <- c("Obs1", "Obs2", "Obs3", "Obs4")

## Load Original KMeans Cluster Assignments

# Load the original cluster assignments from the catch22 features
catch22_kmeans_orig <- readRDS("Results/Clustering/KMeans/kmeans_catch22.RData")

# Grab the centers from these clusters and add row names to the data frame of centers
catch22_kmeans_orig_cent <- catch22_kmeans_orig$centers
row.names(catch22_kmeans_orig_cent) <- c("Cent1", "Cent2", "Cent3", "Cent4", "Cent5")

catch22_kmeans_orig_cent

# Compute the Euclidean distance between each of our new time series and each cluster center
as.matrix(dist(rbind(catch22_kmeans_orig_cent, train_catch22_feat_df_scaled)))[6:9, 1:5]

# Assigned each new time series to the closest center
catch22_clust_assign <- c(1, 1, 4, 4)

# TSFeat Based Feature Computation

# Compute basic set of tsfeatures
# Set scale to FALSE as to not scale the time series before computing the features
ts_feat_list <- tsfeatures(train_samples_ts, scale=FALSE)

# Remove features with very low variation (these were chosen in notebook 7)
ts_feat_list <- ts_feat_list%>% select(-c(frequency, nperiods, seasonal_period1, seasonal_period2))

## Read in Original TSFeat Features

# Read in the TS features computed on the original training data
tsfeat_orig <- read_csv("Results/Clustering/KMeans/tsfeat_feat_df.csv")[,-1]

### Scale with Original Data

# Create a df to save scaled ts features
ts_feat_list_df_scaled <- ts_feat_list

# We use a min max scaling method, with the original data used for scaling the new data
# Loop through each feature
for (n in 1:ncol(ts_feat_list_df_scaled)){
    # Copmute the min and max of each feature from the original data
    min_n <- min(tsfeat_orig[,n])
    max_n <- max(tsfeat_orig[,n])
    # Scaled the new data using the min and max from the original data
    ts_feat_list_df_scaled[,n] <- (ts_feat_list_df_scaled[,n] - min_n)/(max_n - min_n)
}

# Add row names to the data frame
row.names(ts_feat_list_df_scaled) <- c("Obs1", "Obs2", "Obs3", "Obs4")

## Load Original KMeans Cluster Assignments

# Read in the original cluster assignments
tsfeat_kmeans_orig <- readRDS("Results/Clustering/KMeans/kmeans_tsfeat.RData")

# Grab the centers for each cluster and add row names to the data frame of centers
tsfeat_kmeans_orig_cent <- tsfeat_kmeans_orig$centers
row.names(tsfeat_kmeans_orig_cent) <- c("Cent1", "Cent2")

# Compute the Euclidean distance from each new observation to each center
as.matrix(dist(rbind(tsfeat_kmeans_orig_cent, ts_feat_list_df_scaled)))[3:6, 1:2]

# Create a list of cluster assignments for the new data based on the distance to the cluster centers
tsfeat_clust_assign <- c(1, 1, 1, 1)

# DTW Clustering

## Read in Original Clustering Results

# Read in the original cluster results for k medoids dtw clusters
dtw_orig <- readRDS("Results/Clustering/DTW/dtw_clust.RData")

dtw_orig

# Extract the centroids from the clusters
dtw_orig_cent <- dtw_orig@centroids

### Read in Original Data to Determine which TS are the Centroids

# Read in all files from the Data/Processed directory - these are for the original sensors, not the unseen sensors
england_fnames <- list.files("Data/Processed/Highways_England/", pattern="*.csv", full.names=TRUE)
england_df_list <- lapply(england_fnames, read_csv)

portland_fnames <- list.files("Data/Processed/Portland/", pattern="*.csv", full.names=TRUE)
portland_df_list <- lapply(portland_fnames, read_csv)

utah_fnames <- list.files("Data/Processed/Utah/", pattern="*.csv", full.names=TRUE)
utah_df_list <- lapply(utah_fnames, read_csv)

# Ensure that the rows in each data frame are in proper chronological order 
england_df_list <- lapply(england_df_list, function(x) x %>% arrange(timestamp))
portland_df_list <- lapply(portland_df_list, function(x) x %>% arrange(timestamp))
utah_df_list <- lapply(utah_df_list, function(x) x %>% arrange(timestamp))    

# Create one list of all data frames
total_df_list_orig <- append(append(england_df_list, portland_df_list), utah_df_list)

# Read in the start and end points for each original time series from csv
start_end_orig <- read_csv("start_end_points.csv")
starting_points <- start_end_orig$start
ending_points <- start_end_orig$end

# For each df in our list, create a row number column called 'rn' - this will allow us to find the starting point
# for each sample using the above starting_points array
total_df_list_orig <- lapply(total_df_list_orig, function(x) x %>% mutate(rn = row_number()))

# Sample each df in the list according to its corresponding starting and ending point - this creates 12-week
# long samples of each data frame which will be used for modeling and testing
total_df_list_orig_samples <- lapply(1:length(total_df_list_orig), 
                                function(x) total_df_list_orig[[x]] %>% 
                                    filter(rn >= starting_points[x]) %>%
                                    filter(rn <= ending_points[x]))

# Add a new column to each data frame to denote whether each row should be part of the training set, validation set
# or test set - the first 8 weeks are designated as train, the next 2 as val, and the final 2 as test. It may be 
# the case that train and val get used for training, depending on the modeling task
total_df_list_orig_samples <- lapply(total_df_list_orig_samples, 
                                function(x) x %>% mutate(rn = row_number()) %>%
                                    mutate(train_val_test = ifelse(rn <= (96*7*8), 
                                                                   "train", 
                                                                   ifelse(rn <= (96*7*10), 
                                                                          "val", 
                                                                          "test")))
                                
                               )

# Create a list of data frames which only have the training rows of each df
train_samples_orig <- lapply(total_df_list_orig_samples, function(x) x %>% filter(train_val_test == "train"))

# Get the target field (total_volume) from the dfs in the above lists and create msts objects instead of dfs
train_samples_ts_orig <- lapply(train_samples_orig, 
                           function(x) msts(x$total_volume, seasonal.periods=c(24*4, 24*4*7)))

# Loop through the time series in the list
for (i in 1:length(train_samples_ts_orig)) {
    # If the time series is equal to a cluster centroid
    if (mean(train_samples_ts_orig[[i]] == dtw_orig_cent[[1]])==1) {
        # Assign the centroid variable to the time series index
        centroid_1 <- i
    }
    if (mean(train_samples_ts_orig[[i]] == dtw_orig_cent[[2]])==1) {
        centroid_2 <- i
    }
    
}

# Print the index of each of the cluster centroids
centroid_1

centroid_2

## Compute Distance to Centroids

# Compute the DTW distance of each new time series to the cluster centroid for both of the original clusters
# We use the dtw_basic function to match the way distance was computed when the clusters were created
dist_obs_1_cent_1 <- dtw_basic(train_samples_ts[[1]], train_samples_ts_orig[[centroid_1]])
dist_obs_1_cent_2 <- dtw_basic(train_samples_ts[[1]], train_samples_ts_orig[[centroid_2]])

dist_obs_2_cent_1 <- dtw_basic(train_samples_ts[[2]], train_samples_ts_orig[[centroid_1]])
dist_obs_2_cent_2 <- dtw_basic(train_samples_ts[[2]], train_samples_ts_orig[[centroid_2]])

dist_obs_3_cent_1 <- dtw_basic(train_samples_ts[[3]], train_samples_ts_orig[[centroid_1]])
dist_obs_3_cent_2 <- dtw_basic(train_samples_ts[[3]], train_samples_ts_orig[[centroid_2]])

dist_obs_4_cent_1 <- dtw_basic(train_samples_ts[[4]], train_samples_ts_orig[[centroid_1]])
dist_obs_4_cent_2 <- dtw_basic(train_samples_ts[[4]], train_samples_ts_orig[[centroid_2]])

### Observation 1

# Print the distance of each observation to each cluster centroid
dist_obs_1_cent_1

dist_obs_1_cent_2

min(dist_obs_1_cent_1, dist_obs_1_cent_2)

### Observation 2

dist_obs_2_cent_1

dist_obs_2_cent_2

min(dist_obs_2_cent_1, dist_obs_2_cent_2)

### Observation 3

dist_obs_3_cent_1

dist_obs_3_cent_2

min(dist_obs_3_cent_1, dist_obs_3_cent_2)

### Observation 4

dist_obs_4_cent_1

dist_obs_4_cent_2

min(dist_obs_4_cent_1, dist_obs_4_cent_2)

# Based on the printed results, create a list of cluster assignments (whichever centroid was closer)
dtw_clust_assign <- c(2, 2, 1, 1)

# Save Results

# Create a data frame of cluster assignments for each clustering method
new_clust_assign_df <- data.frame(rand=rand_clust_assign,
                                  catch22=catch22_clust_assign,
                                  tsfeat=tsfeat_clust_assign,
                                  dtw=dtw_clust_assign)

# Save the data frame to a csv file
write.csv(new_clust_assign_df, "Results/Unseen Sensor/clust_assign.csv")
