==> 14. Unseen Sensors Forecast with R-Based Models.ipynb <==
# Clear the work space and then garbage collect. We do this in (almost) every notebook to free up memory
rm(list=ls())

gc()

# Uncomment and run to install packages if needed
# install.packages("tidyverse") 
# install.packages("cluster")
# install.packages("forecast")
# install.packages("Metrics")
# install.packages("biglm")
# install.packages("tseries")
# install.packages("scoringutils")

# Load libraries
library(tidyverse)
library(parallel)
library(cluster)
library(forecast)
library(Metrics)
library(biglm)
library(tseries)
library(scoringutils)

# Read In Data

# Read in the cluster assignments for the new sensors (developed in notebook 13)
clust_assign <- read_csv("Results/Unseen Sensor/clust_assign.csv")

# Read in the unseen sensor data
fnames <- list.files("Data/Unseen Sensor/Processed/", pattern="*.csv", full.names=TRUE)
total_df_list <- lapply(fnames, read_csv)

# Ensure all time series are in proper timestamp order
total_df_list <- lapply(total_df_list, function(x) x %>% arrange(timestamp))    

# Read in the start and end points for each time series sample from csv
start_end <- read_csv("start_end_points_unseen.csv")
starting_points <- start_end$start
ending_points <- start_end$end

# For each df in our list, create a row number column called 'rn' - this will allow us to find the starting point
# for each sample using the above starting_points array
total_df_list <- lapply(total_df_list, function(x) x %>% mutate(rn = row_number()))

# Sample each df in the list according to its corresponding starting and ending point - this creates 12-week
# long samples of each data frame which will be used for modeling and testing
total_df_list_samples <- lapply(1:length(total_df_list), 
                                function(x) total_df_list[[x]] %>% 
                                    filter(rn >= starting_points[x]) %>%
                                    filter(rn <= ending_points[x]))

# Add a new column to each data frame to denote whether each row should be part of the training set, validation set
# or test set - the first 8 weeks are designated as train, the next 2 as val, and the final 2 as test. It may be 
# the case that train and val get used for training, depending on the modeling task
total_df_list_samples <- lapply(total_df_list_samples, 
                                function(x) x %>% mutate(rn = row_number()) %>%
                                    mutate(train_val_test = ifelse(rn <= (96*7*8), 
                                                                   "train", 
                                                                   ifelse(rn <= (96*7*10), 
                                                                          "val", 
                                                                          "test")))
                                
                               )

# Set the seed
set.seed(54321)

# Functions for Global AR
We are essentially copying and pasting the functions from notebook number 8. Ideally, we would create these functions outside of either notebook and import them. This would be a better structure and make it easier to maintain or update the logic

lag_embed <- function(data, lag=1) {
    # This function creates a lag emebedded matrix at the provided lag level. It assumes there is a train_val_test
    # in the input data frame, and returns a list of the training, validaiton, and test data as lag matrices
    
    # Compute the number of rows of test and validation data
    n_test <- nrow(data %>% filter(train_val_test=="test"))
    n_val <- nrow(data %>% filter(train_val_test=="val"))
    
    # Lag embed the entire time series - note we use the input lag+1 to account for the target variable
    embed_mat <- embed(data$total_volume, (lag+1))

    # Get the total numebr of rows in the lag matrix
    n_mat <- nrow(embed_mat)
    
    # Based on the number of rows in the matrix and the number of test and validation samples,
    # compute the indexes which correspond to test and validation
    test_idx <- rev(1:n_mat)[1:n_test]
    val_idx <- rev(1:n_mat)[-c(1:n_test)][1:n_val]
    
    # Remoe the test and val indexes to get the training matrix
    train_mat <- embed_mat[-c(val_idx, test_idx),]
    
    # Filter using the above computed indexes to get the validaiton and test matrices
    val_mat <- embed_mat[val_idx,]
    test_mat <- embed_mat[test_idx,]
    
    # Return a list of the trianing, validation, and test lag-embedded matrices
    return_list <- list("train"=train_mat, "val"=val_mat, "test"=test_mat)
}

global_ar_least_squares_pred <- function(model,
                                         test_list,
                                         return_preds=TRUE,
                                         return_means=TRUE
                                        ) 
{
    # Function which takes in a model and a list of test matrices and makes predictions 
    # and computes performance measurements on that test data
    
    # Rename the model object
    big_lm_mod <- model
    
    # Initialize an empty df for storing the predictions
    test_preds <- data.frame()
    # For each matrix in the test list
    for (test_t in 1:length(test_list)) {
        # Compute the one-step ahead predictions
        test_pred <- data.frame(forec=predict(big_lm_mod, newdata=test_list[[test_t]]), ts_index=test_t)
        # Append those to the test_pred df
        test_preds <- rbind(test_preds, test_pred)
    }

    # Initialize empty lists for rmse, mae, and means
    test_rmse_ls <- c()
    test_mae_ls <- c()
    test_means <- c()
    
    # Loop through each time series in the data set and
    for (idx in 1:length(test_list)) {
        
        # Compute the mean and append to the list of means
        test_mean_i <- mean(test_list[[idx]][,1])
        test_means <- c(test_means, test_mean_i)
        
        # Compute the prediction RMSE and append to the list of RMSE
        rmse_test_i <- rmse(test_list[[idx]][,1], (test_preds %>% filter(ts_index==idx))$forec)
        test_rmse_ls <- c(test_rmse_ls, rmse_test_i)
    
        # Copmute the MAE of the predictions and append to list
        mae_test_i <- mae(test_list[[idx]][,1], (test_preds %>% filter(ts_index==idx))$forec)
        test_mae_ls <- c(test_mae_ls, mae_test_i)
    }
    
    # Set up the return_list to include the rmse and mae lists
    return_list <- list("test_rmse"=test_rmse_ls, "test_mae"=test_mae_ls)
    
    # If the user specifies, return the test set predicitons and the target means as well
    if (return_preds) {
        return_list <- append(return_list, list("test_preds"=test_preds))
    }
    
    if (return_means) {
        return_list <- append(return_list, list("test_means"=test_means))
    }
    
    # Return the return_list
    return_list        
}

bootstrap_pred_int <- function(predictions,
                               residuals,
                               n_boot,
                               actuals=NULL
                              ) 
{
    # Function to compute prediction intervals using bootstrapping with model residuals
    # function takes in the predicted values, the model residuals, the number of bootstrap samples to use,
    # and an optional set of true values (used for computing prediction interval scores)
    
    # Intialize empty lists for the hi and lo components of the 80% nad 95% PIs
    boot_lo_80 <- c()
    boot_hi_80 <- c()
    boot_lo_95 <- c()
    boot_hi_95 <- c()

    # Loop through every row in the predictions
    for (i in 1:nrow(predictions)){
        # Reset the seed to the loop counter - could make this more random, but this is for reproducibility
        set.seed(i)
        
        # Sample with replacement from the residuals n_boot times
        boot_samples <- sample(residuals, n_boot, replace=TRUE)
        # Add the predicted value to each sample
        boot_preds <- boot_samples + predictions$forec[i]
        
        # Compute the PIs as the quantiles of these pred+boot values
        boot_pred_lo_80 <- quantile(boot_preds, 0.1)
        boot_pred_hi_80 <- quantile(boot_preds, 0.9)

        boot_pred_lo_95 <- quantile(boot_preds, 0.025)
        boot_pred_hi_95 <- quantile(boot_preds, 0.975)
        
        # Append the PI values to the lists
        boot_lo_80 <- c(boot_lo_80, boot_pred_lo_80)
        boot_hi_80 <- c(boot_hi_80, boot_pred_hi_80)

        boot_lo_95 <- c(boot_lo_95, boot_pred_lo_95)
        boot_hi_95 <- c(boot_hi_95, boot_pred_hi_95)        
    }
    
    # Create a copy of the input df called predicitons
    pred_w_boot_int_df <- predictions
    
    # Add columns to the df for the PIs
    pred_w_boot_int_df$lo_80 <- boot_lo_80
    pred_w_boot_int_df$hi_80 <- boot_hi_80
    pred_w_boot_int_df$lo_95 <- boot_lo_95
    pred_w_boot_int_df$hi_95 <- boot_hi_95
    
    # If the true values are included
    if (!is.null(actuals)){
        
        # Add a column for the true values
        pred_w_boot_int_df$true_values <- actuals
        
        # Compute the 80% and 95% interval scores and add them as columns in the df
        pred_w_boot_int_df <- pred_w_boot_int_df %>%
            mutate(int_score_80 = interval_score(true_values=true_values,
                                                 lower=lo_80,
                                                 upper=hi_80,
                                                 interval_range=80,
                                                 weigh=FALSE
                                                )) %>% 
            mutate(int_score_95 = interval_score(true_values=true_values,
                                                 lower=lo_95,
                                                 upper=hi_95,
                                                 interval_range=95,
                                                 weigh=FALSE
                                                ))
    }
    
    # Return the final set of preds with PIs (and interval scores if applicable)
    pred_w_boot_int_df
    
}

# Global AR Model for Full Data Set

# Create the final lag embedd matrix list using the lag of 840
embed_mat_list <- lapply(total_df_list_samples,
                         function(x) lag_embed(x, lag=840))

# Since we are training the full model now, we will combine the train and validation data to fit the model and only
# test on the test data - thus, we split these up into two separate lists of lag matrices
embed_train_val_list <- lapply(embed_mat_list,
                               function(x) rbind(data.frame(x$train), data.frame(x$val)))
 
embed_test_list <- lapply(embed_mat_list,
                          function(x) data.frame(x$test))   

# Read in the full model which was developed on all of the original data
full_model <- readRDS("Results/Global/Pooled Regression/Full/global_ar_model.RData")

# Make predicitons on the test data using this model
global_ar_full <- global_ar_least_squares_pred(full_model$model, embed_test_list, TRUE, TRUE)

# Create a data frame of the test data
test_data_full <- do.call("rbind", embed_test_list)

# Create bootstrap pred intervals on the test data, using a bootstrap sample number of n_samp=1000
# We could use a smaller n_samp for speed, but higher will give us a more accurate picture of the PIs
n_samp <- 1000
boot_test <- bootstrap_pred_int(global_ar_full$test_preds,
                                full_model$train_residuals,
                                n_samp,
                                test_data_full[,1])     

# Examine the model performance
mean(global_ar_full$test_rmse)

mean(global_ar_full$test_mae)

# Examine scaled model performance
mean(global_ar_full$test_rmse/global_ar_full$test_means)

mean(global_ar_full$test_mae/global_ar_full$test_means)

# Examine 80% PI and scaled 80% PI
mean((boot_test %>% group_by(ts_index) %>% summarise(int_score_80=mean(int_score_80)))$int_score_80)

mean((boot_test %>% group_by(ts_index) %>% summarise(int_score_80=mean(int_score_80)))$int_score_80/
global_ar_full$test_means)

# Examine 95% PI and scaled 95% PI
mean((boot_test %>% group_by(ts_index) %>% summarise(int_score_95=mean(int_score_95)))$int_score_95)

mean((boot_test %>% group_by(ts_index) %>% summarise(int_score_95=mean(int_score_95)))$int_score_95/
global_ar_full$test_means)

# Global AR Model for Highway System

# Load in the model object for the highways england data
highways_england_model <- readRDS("Results/Global/Pooled Regression/Highway System/global_ar_model_country_1.RData")

# Make predicitons on the test data using this model
global_ar_england_pred <- global_ar_least_squares_pred(highways_england_model$model, embed_test_list, TRUE, TRUE)

# Create bootstrap pred intervals on the test data, using a bootstrap sample number of n_samp=1000
# We could use a smaller n_samp for speed, but higher will give us a more accurate picture of the PIs
n_samp <- 1000
boot_test_england <- bootstrap_pred_int(global_ar_england_pred$test_preds,
                                        highways_england_model$train_residuals,
                                        n_samp,
                                        test_data_full[,1])    

# Examine the model performance
mean(global_ar_england_pred$test_rmse)

mean(global_ar_england_pred$test_mae)

mean(global_ar_england_pred$test_rmse/global_ar_england_pred$test_means)

mean(global_ar_england_pred$test_mae/global_ar_england_pred$test_means)

# Examine model PI performance
mean((boot_test_england %>% group_by(ts_index) %>% summarise(int_score_80=mean(int_score_80)))$int_score_80)

mean((boot_test_england %>% group_by(ts_index) %>% summarise(int_score_80=mean(int_score_80)))$int_score_80/
global_ar_england_pred$test_means)

mean((boot_test_england %>% group_by(ts_index) %>% summarise(int_score_95=mean(int_score_95)))$int_score_95)

mean((boot_test_england %>% group_by(ts_index) %>% summarise(int_score_95=mean(int_score_95)))$int_score_95/
global_ar_england_pred$test_means)

# Global AR Random Clustering

# Print out the cluster assignemnts. 
# We will load the models in the order of the cluster assignments to simplify things
clust_assign$rand

# Load in the models in the order of the cluster assignments (i.e., read in model 3 first because the first time
# series in our data set is part of cluster 3, and so on)
rand_clust_mod3 <- readRDS("Results/Global/Pooled Regression/Random Cluster/global_ar_model_cluster_3.RData")
rand_clust_mod2 <- readRDS("Results/Global/Pooled Regression/Random Cluster/global_ar_model_cluster_2.RData")
rand_clust_mod1 <- readRDS("Results/Global/Pooled Regression/Random Cluster/global_ar_model_cluster_1.RData")
rand_clust_mod4 <- readRDS("Results/Global/Pooled Regression/Random Cluster/global_ar_model_cluster_4.RData")

# Make predicitons on the test data for each cluster using each 
test_pred_clust3 <- global_ar_least_squares_pred(rand_clust_mod3$model, embed_test_list[1], TRUE, TRUE)

test_pred_clust2 <- global_ar_least_squares_pred(rand_clust_mod2$model, embed_test_list[2], TRUE, TRUE)

test_pred_clust1 <- global_ar_least_squares_pred(rand_clust_mod1$model, embed_test_list[3], TRUE, TRUE)

test_pred_clust4 <- global_ar_least_squares_pred(rand_clust_mod4$model, embed_test_list[4], TRUE, TRUE)

# Create bootstrap pred intervals on the test data, using a bootstrap sample number of n_samp=1000
n_samp <- 1000
boot_testclust3 <- bootstrap_pred_int(test_pred_clust3$test_preds,
                                      rand_clust_mod3$train_residuals,
                                      n_samp,
                                      embed_test_list[[1]][,1]) 

boot_testclust2 <- bootstrap_pred_int(test_pred_clust2$test_preds,
                                      rand_clust_mod2$train_residuals,
                                      n_samp,
                                      embed_test_list[[2]][,1]) 

boot_testclust1 <- bootstrap_pred_int(test_pred_clust1$test_preds,
                                      rand_clust_mod1$train_residuals,
                                      n_samp,
                                      embed_test_list[[3]][,1]) 

boot_testclust4 <- bootstrap_pred_int(test_pred_clust4$test_preds,
                                      rand_clust_mod4$train_residuals,
                                      n_samp,
                                      embed_test_list[[4]][,1]) 

# Compute the per time series mean of the forecast RMSE
mean(c(test_pred_clust3$test_rmse, test_pred_clust2$test_rmse, 
       test_pred_clust1$test_rmse, test_pred_clust4$test_rmse))

# Compute the per time series mean of the forecast MAE
mean(c(test_pred_clust3$test_mae, test_pred_clust2$test_mae,
       test_pred_clust1$test_mae, test_pred_clust4$test_mae))

# Compute the scaled per time series mean of the forecast RMSE
mean(c(test_pred_clust3$test_rmse, test_pred_clust2$test_rmse, test_pred_clust1$test_rmse, test_pred_clust4$test_rmse)/
     c(test_pred_clust3$test_means, test_pred_clust2$test_means, test_pred_clust1$test_means, test_pred_clust4$test_means))

# Compute the scaled per time series mean of the forecast MAE
mean(c(test_pred_clust3$test_mae, test_pred_clust2$test_mae, test_pred_clust1$test_mae, test_pred_clust4$test_mae)/
     c(test_pred_clust3$test_means, test_pred_clust2$test_means, test_pred_clust1$test_means, test_pred_clust4$test_means))

# Compute the mean interval score at 80%
mean((rbind(boot_testclust3, boot_testclust2, boot_testclust1, boot_testclust4))$int_score_80)

# Compute the mean scaled interval score at 80%
mean(c(mean(boot_testclust3$int_score_80), mean(boot_testclust2$int_score_80), 
mean(boot_testclust1$int_score_80), mean(boot_testclust4$int_score_80))/
c(test_pred_clust3$test_means, test_pred_clust2$test_means, 
  test_pred_clust1$test_means, test_pred_clust4$test_means))

# Compute the scaled interval score at 95%
mean((rbind(boot_testclust3, boot_testclust2, boot_testclust1, boot_testclust4))$int_score_95)

# Compute the mean scaled interval score at 95%
mean(c(mean(boot_testclust3$int_score_95), mean(boot_testclust2$int_score_95), 
mean(boot_testclust1$int_score_95), mean(boot_testclust4$int_score_95))/
c(test_pred_clust3$test_means, test_pred_clust2$test_means, 
  test_pred_clust1$test_means, test_pred_clust4$test_means))

# Global AR Catch22

# Print the Catch22 cluster assignments
clust_assign$catch22

# Load models for cluster numbers 1 and 4
catch22_mod1 <- readRDS("Results/Global/Pooled Regression/Catch22 KMeans/global_ar_model_kmeans_catch22_clust_1.RData")

catch22_mod4 <- readRDS("Results/Global/Pooled Regression/Catch22 KMeans/global_ar_model_kmeans_catch22_clust_4.RData")

# Make predicitons on the subsets of the test data for each cluster and model
test_pred_catch22_clust1 <- global_ar_least_squares_pred(catch22_mod1$model, 
                                                         embed_test_list[1:2], TRUE, TRUE)

test_pred_catch22_clust4 <- global_ar_least_squares_pred(catch22_mod4$model, 
                                                         embed_test_list[3:4], TRUE, TRUE)

# Create bootstrap pred intervals on the test data, using a bootstrap sample number of n_samp=1000
n_samp <- 1000
boot_catch22_testclust1 <- bootstrap_pred_int(test_pred_catch22_clust1$test_preds,
                                              catch22_mod1$train_residuals,
                                              n_samp,
                                              do.call(rbind, embed_test_list[1:2])[,1])

boot_catch22_testclust4 <- bootstrap_pred_int(test_pred_catch22_clust4$test_preds,
                                              catch22_mod4$train_residuals,
                                              n_samp,
                                              do.call(rbind, embed_test_list[3:4])[,1])

# Inspect model performance in the next several cells
mean(c(test_pred_catch22_clust1$test_rmse, test_pred_catch22_clust4$test_rmse))

mean(c(test_pred_catch22_clust1$test_mae,test_pred_catch22_clust4$test_mae))

mean(c(test_pred_catch22_clust1$test_rmse, test_pred_catch22_clust4$test_rmse)/
     c(test_pred_catch22_clust1$test_means, test_pred_catch22_clust4$test_means)
     )

mean(c(test_pred_catch22_clust1$test_mae, test_pred_catch22_clust4$test_mae)/
     c(test_pred_catch22_clust1$test_means, test_pred_catch22_clust4$test_means)
     )

mean(c(boot_catch22_testclust1$int_score_80, boot_catch22_testclust4$int_score_80))

# To compute the scaled PI performance, we group each clustered set of PIs by time series index and compute the mean
# and then divide the true means. We do this because the time series index returned by the modeling/test set prediction
# functions is not currently correct
mean(c((boot_catch22_testclust1 %>% group_by(ts_index) %>% summarise(int_score_80=mean(int_score_80)))$int_score_80,
(boot_catch22_testclust4 %>% group_by(ts_index) %>% summarise(int_score_80=mean(int_score_80)))$int_score_80)/
c(test_pred_catch22_clust1$test_means, test_pred_catch22_clust4$test_means)
     )

mean(c(boot_catch22_testclust1$int_score_95, boot_catch22_testclust4$int_score_95))

mean(c((boot_catch22_testclust1 %>% group_by(ts_index) %>% summarise(int_score_95=mean(int_score_95)))$int_score_95,
(boot_catch22_testclust4 %>% group_by(ts_index) %>% summarise(int_score_95=mean(int_score_95)))$int_score_95)/
c(test_pred_catch22_clust1$test_means, test_pred_catch22_clust4$test_means)
     )

# GlobalAR TSFeat

# All time series are grouped into cluster one for this feature set
clust_assign$tsfeat

# Load the model for cluster 1
tsfeat_mod1 <- readRDS("Results/Global/Pooled Regression/TSFeat KMeans/global_ar_model_kmeans_tsfeat_clust_1.RData")

# Make predicitons on the test data using this model
test_pred_tsfeat_clust1 <- global_ar_least_squares_pred(tsfeat_mod1$model, 
                                                        embed_test_list, TRUE, TRUE)

# Create bootstrap pred intervals on the test data, using a bootstrap sample number of n_samp=1000
n_samp <- 1000
boot_tsfeat_testclust1 <- bootstrap_pred_int(test_pred_tsfeat_clust1$test_preds,
                                             tsfeat_mod1$train_residuals,
                                             n_samp,
                                             test_data_full[,1])

# Print model performance metrics
mean(test_pred_tsfeat_clust1$test_rmse)

mean(test_pred_tsfeat_clust1$test_mae)

mean(test_pred_tsfeat_clust1$test_rmse/
    test_pred_tsfeat_clust1$test_means)

mean(test_pred_tsfeat_clust1$test_mae/
    test_pred_tsfeat_clust1$test_means)

# Print PI related metrics
mean(boot_tsfeat_testclust1$int_score_80)

mean((boot_tsfeat_testclust1 %>% group_by(ts_index) %>% summarise(int_score_80=mean(int_score_80)))$int_score_80/
test_pred_tsfeat_clust1$test_means)

mean(boot_tsfeat_testclust1$int_score_95)

mean((boot_tsfeat_testclust1 %>% group_by(ts_index) %>% summarise(int_score_95=mean(int_score_95)))$int_score_95/
test_pred_tsfeat_clust1$test_means)

# GlobalAR DTW

# Print the cluster assignments 
clust_assign$dtw

# Load the models for the clustered data
dtw_clust_mod2 <- readRDS("Results/Global/Pooled Regression/DTW/global_ar_dtw_clust_2.RData")
dtw_clust_mod1 <- readRDS("Results/Global/Pooled Regression/DTW/global_ar_dtw_clust_1.RData")

# Make predicitons on the clustered test data using these model models
test_pred_dtw_clust2 <- global_ar_least_squares_pred(dtw_clust_mod2$model,
                                                     embed_test_list[1:2], TRUE, TRUE)

test_pred_dtw_clust1 <- global_ar_least_squares_pred(dtw_clust_mod1$model,
                                                     embed_test_list[3:4], TRUE, TRUE)

# Create bootstrap pred intervals on the clustered test data, using a bootstrap sample number of n_samp=1000
n_samp <- 1000

boot_dtwclust_2 <- bootstrap_pred_int(test_pred_dtw_clust2$test_preds,
                                      dtw_clust_mod2$train_residuals,
                                      n_samp,
                                      do.call(rbind, embed_test_list[1:2])[,1]) 


boot_dtwclust_1 <- bootstrap_pred_int(test_pred_dtw_clust1$test_preds,
                                      dtw_clust_mod1$train_residuals,
                                      n_samp,
                                      do.call(rbind, embed_test_list[3:4])[,1]) 

# Create a list of RMSE and MAE and inspect model performance
mean(c(test_pred_dtw_clust1$test_rmse, test_pred_dtw_clust2$test_rmse))

mean(c(test_pred_dtw_clust1$test_mae, test_pred_dtw_clust2$test_mae))

mean(c(test_pred_dtw_clust1$test_rmse, test_pred_dtw_clust2$test_rmse)/
     c(test_pred_dtw_clust1$test_means, test_pred_dtw_clust2$test_means)
    )

mean(c(test_pred_dtw_clust1$test_mae, test_pred_dtw_clust2$test_mae)/
     c(test_pred_dtw_clust1$test_means, test_pred_dtw_clust2$test_means)
    )

# Inspect prediction interval performance
mean((rbind(boot_dtwclust_1, boot_dtwclust_2))$int_score_80)

mean(c((boot_dtwclust_1 %>% group_by(ts_index) %>% summarise(int_score_80=mean(int_score_80)))$int_score_80,
(boot_dtwclust_2 %>% group_by(ts_index) %>% summarise(int_score_80=mean(int_score_80)))$int_score_80)/
c(test_pred_dtw_clust1$test_means, test_pred_dtw_clust2$test_means))

mean((rbind(boot_dtwclust_1, boot_dtwclust_2))$int_score_95)

mean(c((boot_dtwclust_1 %>% group_by(ts_index) %>% summarise(int_score_95=mean(int_score_95)))$int_score_95,
(boot_dtwclust_2 %>% group_by(ts_index) %>% summarise(int_score_95=mean(int_score_95)))$int_score_95)/
c(test_pred_dtw_clust1$test_means, test_pred_dtw_clust2$test_means))
