==> 15. Unseen Sensor LightGBM Default Forecasts.ipynb <==
# If the libraries are not yet installed, they can be installed in this notebook using commands similar to the below
# %conda install numpy
# %conda install pandas
# %conda install matplotlib
# %conda install scikit-learn
# %conda install -c conda-forge lightgbm 
# %conda install -c conda-forge swifter
# %conda install -c conda-forge scipy
# %conda install joblib
# %conda install tdqm

# Something like the following may also work if the above does not
# import sys
# !conda install --yes --prefix {sys.prefix} numpy
# !conda install --yes --prefix {sys.prefix} pandas
# !conda install --yes --prefix {sys.prefix} scikit-learn
# !conda install -c conda-forge --yes --prefix {sys.prefix} lightgbm
# !conda install -c conda-forge --yes --prefix {sys.prefix} swifter
# !conda install -c conda-forge --yes --prefix {sys.prefix} scipy 
# !conda install --yes --prefix {sys.prefix} joblib
# !conda install --yes --prefix {sys.prefix} tdqm

# To install a specific version, add the version to the install command
# E.g., %conda install numpy=1.20.3

# If all else fails, use pip or follow additional advice such as found at
# https://jakevdp.github.io/blog/2017/12/05/installing-python-packages-from-jupyter/

# If your plan to use pip (especially if you are not working within a specified conda environment), 
# the pip commands might look like:
# pip install numpy
# pip install pandas
# pip install scikit-learn
# pip install lightgbm
# pip install swifter
# pip install scipy
# pip install joblib
# pip install tdqm

# To install a specific version, add the version to the pip install command
# E.g., pip install numpy==1.20.3

# Load libraries
import time
import numpy as np
import pandas as pd
import json
import matplotlib.pyplot as plt
import itertools
from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error
import glob
from lightgbm import LGBMRegressor
import random
from sklearn.model_selection import ParameterSampler
import scipy
import gc
from joblib import Parallel, delayed
import contextlib
import joblib
from tqdm import tqdm

# Set the random seeds
np.random.seed(54321)
random.seed(54321)

# Read in Data and Prepare for Modeling

# Create an empty list to hold the dataframes of highways england data
total_df_list = list()

# Loop through the files, sorted in alphabetical order
# Read them into a df, make sure they are sorted by timestamp, and append to the list
for fname in sorted(glob.glob("Data/Unseen Sensor/Processed/*.csv")):
    print("Reading {}".format(fname))
    df = pd.read_csv(fname) #, parse_dates=['timestamp'], index_col=['timestamp'])
    df = df.sort_values(by="timestamp")
    total_df_list.append(df)

# Read in the start and end points csv, and subtract 1 to deal with index differences between R and python
start_end = pd.read_csv("start_end_points_unseen.csv")
start_end["start"] = start_end["start"] - 1
start_end["end"] = start_end["end"]

start_end.head()

# Create an empty list to hold the subset data frames (those with only 12 weeks of data per highway)
subset_df_list = list()

# For each df in our original total df list
for idx, df in enumerate(total_df_list):
        
    # Filter the timeframe based on the start_end_points csv files
    subset_df = df.iloc[start_end.iloc[idx,0]:start_end.iloc[idx,1], ]\
    .reset_index(drop=True).reset_index(drop=False)\
    .rename(columns={"index":"rn"})
    
    # Create a new field called train_val_test to differentiate each set of data
    subset_df["train_val_test"] = np.where(subset_df["rn"]<(96*7*8),
                                           "train",
                                           np.where(subset_df["rn"]<(96*7*10),
                                                    "val",
                                                    "test"
                                                   )
                                       )
    
    # Append to list
    subset_df_list.append(subset_df)

# Create a list of df's with only fields we need

# Initialize empty list
model_df_list = list()

# For df in subset list
for df in subset_df_list:
       
    # Extract the timestamp, the volume, and the train_val_test assignment
    model_df = df[['timestamp', 'total_volume', "train_val_test"]]\
    .rename(columns={'timestamp':'start', 'total_volume':'target'})
    
    # Append this df to the new list
    model_df_list.append(model_df)

## Create Lag Emebedded Matrices for each TS

# Set our lag value to be 840
lag_n = 840

# # Lag embed the data frames and save to a list
lag_embed_df_list = list()

for df in model_df_list:
    # For each df in our list
    for n in range(1, (lag_n+1)):
        # For each lag level, up to 840
        # Create a new column called target-n
        name = f"target-{n}"
        # Save the target shifted n values into this colume
        df[name] = df['target'].shift(n)
    # Append to list
    lag_embed_df_list.append(df)

# Split the lag embedded list into train, val, and test lists

# First, initialize empty lists for each train, val, and test
train_df_list = list()
val_df_list = list()
test_df_list = list()

for i in range(len(lag_embed_df_list)):
    # For each df in our list
    df = lag_embed_df_list[i].copy()

    # Add a ts_index of i+1 to join with clustering data from R
    df['ts_index'] = i + 1
    
    # Subset into train, val, and test df's based on the train_val_test_field
    train_df = df.query("train_val_test == 'train'").copy()
    val_df = df.query("train_val_test=='val'").copy()
    test_df = df.query("train_val_test=='test'").copy()
    
    # Append to appropriate lists
    train_df_list.append(train_df)
    val_df_list.append(val_df)
    test_df_list.append(test_df)

# Concat all dfs from the lists together to create one full train, val, and test df
train_df_full = pd.concat(train_df_list)
val_df_full = pd.concat(val_df_list)
test_df_full = pd.concat(test_df_list)

# Drop unneeded columns
train_df_full.drop(columns=['start', 'train_val_test'], inplace=True)
val_df_full.drop(columns=['start', 'train_val_test'], inplace=True)
test_df_full.drop(columns=['start', 'train_val_test'], inplace=True)

# Append the training and validation data together for later use
train_val_df_full = train_df_full.append(val_df_full)

# Delete unused variables to free up memory
del train_df_list
del val_df_list 
del test_df_list
del lag_embed_df_list
del model_df_list
del subset_df_list

# Force garbage collection to free up memory
gc.collect()

# Read in the clustering assignment for these new sensors
unseen_sen_clust = pd.read_csv("Results/Unseen Sensor/clust_assign.csv")

unseen_sen_clust.head()

# Full Model

# Use joblib to load the full model
filename = 'Results/Global/LightGBM Default/Full/model'
mod = joblib.load(filename)

# Read in the residuals from the full model as well
res = pd.read_csv('Results/Global/LightGBM Default/Full/residual.csv')

# Function to compute test preds
def compute_lgbm_test_preds(mod, data, lag_n):
    """Function which takes in: a model, test data, and the lag embedding to use, and returns a df of forecasts"""

    # Initialize an empty data frame to store preds
    pred_df = pd.DataFrame()
    
    # Loop through each individual time series index in the data set
    for ts_idx in data.ts_index.unique():
        # Create the X matrix for each one
        X = data.query("ts_index==@ts_idx").iloc[:,1:(lag_n+1)].copy()

        # Forecast for that X matrix
        preds = mod.predict(X)
        
        # Save the results to a temp data frame
        pred_df_sub = pd.DataFrame({"ts_index": ts_idx, "test_preds": preds})
        
        # Append to primary data frame
        pred_df = pred_df.append(pred_df_sub)
    
    # Return df of all preds with corresponding ts_index column
    return pred_df

# Compute the test preds for the unseen sensors with the full model
unseen_test_preds = compute_lgbm_test_preds(mod, test_df_full, lag_n)

# Function to compute performance metrics on test data
def compute_lgbm_test_perf(preds, data):
    """Function which takes inputs: a data frame of test predictions, and a test data df,
    and which returns a data frame of model performance"""
    
    # Create an empty list to store model performance
    perf_ls = list()
    
    # For each time series index in our data set
    for ts_idx in data.ts_index.unique():
        # Get the target (actual) for that index
        y_sub = data.query("ts_index==@ts_idx").iloc[:,0]
        # Extract the corresponding forecasts
        preds_sub = preds.query("ts_index==@ts_idx").test_preds
        
        # Compute rmse, mae, and the mean of the true target value for those preds
        rmse_sub = mean_squared_error(y_sub, preds_sub, squared=False)
        mae_sub = mean_absolute_error(y_sub, preds_sub)
        mean_sub = np.mean(y_sub)
        
        # Save those metrics to a dictionary
        pred_dict = {"rmse": rmse_sub, "mae": mae_sub, "mean": mean_sub}
        
        # Append the dictionary to the list
        perf_ls.append(pred_dict)
        
    # Return a data frame of model performance created from the list of dictionaries
    return pd.DataFrame(perf_ls)

# Compute model perf metrics using above function
full_mod_test_perf = compute_lgbm_test_perf(unseen_test_preds, test_df_full)

# Compute scaled performance metrics in new columns
full_mod_test_perf['nrmse'] = full_mod_test_perf['rmse']/full_mod_test_perf['mean']
full_mod_test_perf['smae'] = full_mod_test_perf['mae']/full_mod_test_perf['mean']

# Print the means of model perf metrics
full_mod_test_perf.mean()

# Function to compute pred intervals with bootstrap method
def compute_lgbm_boostrap_int(preds, resid, n_boot):
    """Function which takes in a model's predictions and residuals, and a number of bootstrap resamples to use,
    and which outputs a df with pred intervals at 80% and 95%"""
    
    # Set seeds for reproducibility
    random.seed(54321)
    np.random.seed(54321)
    
    # Create empty columns in the pred df to store the PIs
    preds['lo_95'] = np.nan
    preds['hi_95'] = np.nan
    preds['lo_80'] = np.nan
    preds['hi_80'] = np.nan
    
    # For each row in the pred df
    for n in range(preds.shape[0]):
        # Sample with replacement n_boot times from the residuals
        resid_boot = np.random.choice(resid, size=n_boot, replace=True)
        # Extract the forecast value for that row
        pred_n = preds.iloc[n, :].test_preds
        # Add the residual vector to the forecast value
        pred_n_boot = resid_boot + pred_n
        
        # Compute quantiles of this residual+forecast vector
        percent_95_lo = np.percentile(pred_n_boot, 2.5)
        percent_95_hi = np.percentile(pred_n_boot, 97.5)
        
        percent_80_lo = np.percentile(pred_n_boot, 10)
        percent_80_hi = np.percentile(pred_n_boot, 90)
        
        # Save these quantiles to the appropriate df column
        preds.iloc[n, 2] = percent_95_lo
        preds.iloc[n, 3] = percent_95_hi
        preds.iloc[n, 4] = percent_80_lo
        preds.iloc[n, 5] = percent_80_hi
    
    # Return the updated preds data frame
    return preds

# Compute PIs with 1000 bootstrap samples
n_boot = 1000

# Here, we pass test preds, the model residuals, and the bootstrappping number to the custom function defined above
# We extract the model residuals from the residual data frame loaded earlier by accessing the residual column and 
# then converting to a list by using the .values method
full_mod_boot_ints = compute_lgbm_boostrap_int(unseen_test_preds, 
                                               res.residual.values, 
                                               n_boot)

# Add the true values into their own df column
full_mod_boot_ints['actual'] = test_df_full.iloc[:,0].to_list()

full_mod_boot_ints.head()

# Create a function to compute the interval score
def interval_score(true_values, lower, upper, interval_range):
    """ Function which takes in the true values, the upper and lower bounds of PIs, and the PI level (e.g., 90%)
        and from these inputs, computes the interval score for each prediction
    """
    
    # Compute alpha from the interval range
    alpha = 1-interval_range
    
    # Save the upper, lower, and true_values as numpy arrays for computation purposes
    upper = np.array(upper)
    lower = np.array(lower)
    true_values = np.array(true_values)
    
    # Compute the lower component of the interval score - just a boolean for true below interval
    def lower_ind(true,low):
        if true<low:
            return 1
        else:
            return 0
        
    # Computer the upper component of the interval score - similar boolean for true above interval
    def upper_ind(true,up):
        if true>up:
            return 1
        else:
            return 0
        
    # Computer the actual score for each obsveration - formula here: https://epiforecasts.io/scoringutils/reference/interval_score.html
    scores = (upper-lower) + (2/alpha)*(lower-true_values)*(lower > true_values) + (2/alpha)*(true_values-upper)*(true_values > upper)
    
    # Return the scores array
    return scores

# Compute the 95% and 80% PI scores using the above function as new data frame columns
full_mod_boot_ints['int_95_score'] = interval_score(full_mod_boot_ints.actual, 
                                                    full_mod_boot_ints.lo_95,
                                                    full_mod_boot_ints.hi_95,
                                                    0.95)
                                                    
full_mod_boot_ints['int_80_score'] = interval_score(full_mod_boot_ints.actual, 
                                                    full_mod_boot_ints.lo_80,
                                                    full_mod_boot_ints.hi_80,
                                                    0.80)

# Print the means of the interval scores
full_mod_boot_ints.mean()

# Group the results by ts_index and compute the scaled interval scores as well

# To do this, we first group by the time series index and compute the mean of the interval scores for each time series
full_mod_boot_ints_group = full_mod_boot_ints.groupby("ts_index").mean().reset_index()
# We then add the true mean of each time series as its own column to the data frame
full_mod_boot_ints_group['mean'] = full_mod_test_perf['mean'].values
# We then scale the interval scores by the true mean
full_mod_boot_ints_group['int_95_score_scaled'] = full_mod_boot_ints_group['int_95_score']/full_mod_boot_ints_group['mean']
full_mod_boot_ints_group['int_80_score_scaled'] = full_mod_boot_ints_group['int_80_score']/full_mod_boot_ints_group['mean']

# Print the average per time series scaled interval score
full_mod_boot_ints_group[['int_95_score_scaled', 'int_80_score_scaled']].mean()

# Random Clusters

# Check the random cluster assignments
unseen_sen_clust.rand

# Load the models for needed to make predictions based on random cluster assignments
rand_clust_mod3 = joblib.load("Results/Global/LightGBM Default/Random Cluster/model_3")
rand_clust_mod2 = joblib.load("Results/Global/LightGBM Default/Random Cluster/model_2")
rand_clust_mod1 = joblib.load("Results/Global/LightGBM Default/Random Cluster/model_1")
rand_clust_mod4 = joblib.load("Results/Global/LightGBM Default/Random Cluster/model_4")

# Load the residual data frame
rand_clust_res = pd.read_csv("Results/Global/LightGBM Default/Random Cluster/residual.csv")

# In this data frame, the list of residuals for each model is stored as an entry in the data frame, and by default
# it is a string, not a true list. We convert each one to a true list with the line of code below
rand_clust_res['residual'] = rand_clust_res['residual'].apply(eval)

# We extract the list of residuals for each model from the data frame and save each in their own variable
res_rand_clust_3 = rand_clust_res.query("cluster==3")['residual'].values[0]
res_rand_clust_2 = rand_clust_res.query("cluster==2")['residual'].values[0]
res_rand_clust_1 = rand_clust_res.query("cluster==1")['residual'].values[0]
res_rand_clust_4 = rand_clust_res.query("cluster==4")['residual'].values[0]

# We compute the test set preds for each model, 
# by subsetting the test df as we pass it to the compute_lgbm_test_preds function
unseen_test_preds_rand_clust3 = compute_lgbm_test_preds(rand_clust_mod3, 
                                                        test_df_full.query("ts_index==1"), 
                                                        lag_n)

unseen_test_preds_rand_clust2 = compute_lgbm_test_preds(rand_clust_mod2, 
                                                        test_df_full.query("ts_index==2"), 
                                                        lag_n)

unseen_test_preds_rand_clust1 = compute_lgbm_test_preds(rand_clust_mod1, 
                                                        test_df_full.query("ts_index==3"), 
                                                        lag_n)

unseen_test_preds_rand_clust4 = compute_lgbm_test_preds(rand_clust_mod4, 
                                                        test_df_full.query("ts_index==4"), 
                                                        lag_n)

# Compute model perf metrics for each clustered result
rand_clust3_test_perf = compute_lgbm_test_perf(unseen_test_preds_rand_clust3, test_df_full.query("ts_index==1"))
rand_clust2_test_perf = compute_lgbm_test_perf(unseen_test_preds_rand_clust2, test_df_full.query("ts_index==2"))
rand_clust1_test_perf = compute_lgbm_test_perf(unseen_test_preds_rand_clust1, test_df_full.query("ts_index==3"))
rand_clust4_test_perf = compute_lgbm_test_perf(unseen_test_preds_rand_clust4, test_df_full.query("ts_index==4"))

# Append all clustered results into one data frame
rand_clust_test_perf = rand_clust3_test_perf.append(rand_clust2_test_perf).append(rand_clust1_test_perf).append(rand_clust4_test_perf)

# Compute scaled performance metrics in new columns
rand_clust_test_perf['nrmse'] = rand_clust_test_perf['rmse']/rand_clust_test_perf['mean']
rand_clust_test_perf['smae'] = rand_clust_test_perf['mae']/rand_clust_test_perf['mean']

# Print the means of the point forecast metrics
rand_clust_test_perf.mean()

# Compute PIs with 1000 bootstrap samples, one function call for each cluster
nboot = 1000

rand3_boot_ints = compute_lgbm_boostrap_int(unseen_test_preds_rand_clust3,
                                            res_rand_clust_3, 
                                            nboot)

rand2_boot_ints = compute_lgbm_boostrap_int(unseen_test_preds_rand_clust2,
                                            res_rand_clust_2, 
                                            nboot)

rand1_boot_ints = compute_lgbm_boostrap_int(unseen_test_preds_rand_clust1,
                                            res_rand_clust_1, 
                                            nboot)

rand4_boot_ints = compute_lgbm_boostrap_int(unseen_test_preds_rand_clust4,
                                            res_rand_clust_4, 
                                            nboot)

# Append all of the clustered boot strap PI data frames into one data frame
rand_boot_ints = rand3_boot_ints.append(rand2_boot_ints).append(rand1_boot_ints).append(rand4_boot_ints)

# Add the true values into their own df column
rand_boot_ints['actual'] = test_df_full.iloc[:,0].to_list()

# Compute the 95% and 80% PI scores using the above function as new data frame columns
rand_boot_ints['int_95_score'] = interval_score(rand_boot_ints.actual, 
                                                rand_boot_ints.lo_95,
                                                rand_boot_ints.hi_95,
                                                0.95)
                                                    
rand_boot_ints['int_80_score'] = interval_score(rand_boot_ints.actual, 
                                                rand_boot_ints.lo_80,
                                                rand_boot_ints.hi_80,
                                                0.80)

# Inspect the PI performance
rand_boot_ints.mean()

# Group by ts_index and compute scaled PI performance
rand_boot_ints_group = rand_boot_ints.groupby("ts_index").mean().reset_index()
rand_boot_ints_group['mean'] = rand_clust_test_perf['mean'].values
rand_boot_ints_group['int_95_score_scaled'] = rand_boot_ints_group['int_95_score']/rand_boot_ints_group['mean']
rand_boot_ints_group['int_80_score_scaled'] = rand_boot_ints_group['int_80_score']/rand_boot_ints_group['mean']

# Print scaled PI performance
rand_boot_ints_group[['int_95_score_scaled', 'int_80_score_scaled']].mean()

# Highway System

# Load the model for highways england data
eng_mod = joblib.load("Results/Global/LightGBM Default/Highway System/model_1")

# Load the residuals, fix the data type
highway_res = pd.read_csv("Results/Global/LightGBM Default/Highway System/residual.csv")
highway_res['residual'] = highway_res['residual'].apply(eval)

# Extract the highways england residuals into a list from the residual data frame
eng_res = highway_res.query("cluster==1")['residual'].values[0]

# Compute test preds with the highways england model
unseen_test_preds_highway = compute_lgbm_test_preds(eng_mod, 
                                                    test_df_full,
                                                    lag_n)

# Compute the performance metrics on the forecasts
highway_test_perf = compute_lgbm_test_perf(unseen_test_preds_highway, test_df_full)


# Compute scaled performance metrics
highway_test_perf['nrmse'] = highway_test_perf['rmse']/highway_test_perf['mean']
highway_test_perf['smae'] = highway_test_perf['mae']/highway_test_perf['mean']

# Print the point forecast metrics
highway_test_perf.mean()

# Compute the prediction intervals for the forecasts
highway_boot_ints = compute_lgbm_boostrap_int(unseen_test_preds_highway,
                                              eng_res,
                                              nboot)

# Add the actual values to the pred int data frame
highway_boot_ints['actual'] = test_df_full.iloc[:,0].to_list()

# Compute the interval scores for each forecast
highway_boot_ints['int_95_score'] = interval_score(highway_boot_ints.actual, 
                                                   highway_boot_ints.lo_95,
                                                   highway_boot_ints.hi_95,
                                                   0.95)
                                                    
highway_boot_ints['int_80_score'] = interval_score(highway_boot_ints.actual, 
                                                   highway_boot_ints.lo_80,
                                                   highway_boot_ints.hi_80,
                                                   0.80)

# Print the mean of the interval scores
highway_boot_ints.mean()

# Group by ts_index and compute the scaled interval scores
highway_boot_ints_group = highway_boot_ints.groupby("ts_index").mean().reset_index()
highway_boot_ints_group['mean'] = highway_test_perf['mean'].values
highway_boot_ints_group['int_95_score_scaled'] = highway_boot_ints_group['int_95_score']/highway_boot_ints_group['mean']
highway_boot_ints_group['int_80_score_scaled'] = highway_boot_ints_group['int_80_score']/highway_boot_ints_group['mean']

# Print the scaled interval scores
highway_boot_ints_group[['int_95_score_scaled', 'int_80_score_scaled']].mean()

# Catch22 KMeans Clusters

# Check the cluster assignments
unseen_sen_clust.catch22

# Load the models needed based on the cluster assignments
catch22_mod1 = joblib.load("Results/Global/LightGBM Default/Catch22 KMeans/model_1")

catch22_mod4 = joblib.load("Results/Global/LightGBM Default/Catch22 KMeans/model_4")

# Read in the residuals for the Catch22 models and cast to proper data types
catch22_res = pd.read_csv("Results/Global/LightGBM Default/Catch22 KMeans/residual.csv")
catch22_res['residual'] = catch22_res['residual'].apply(eval)

# Extract each models residuals from the data frame into its own list
catch22_res1 = catch22_res.query("cluster==1")['residual'].values[0]
catch22_res4 = catch22_res.query("cluster==4")['residual'].values[0]

# Compute the preds for each cluster/model, subsetting the test_df to the proper indexes in the function call
unseen_test_preds_catch22_clust1 = compute_lgbm_test_preds(catch22_mod1, 
                                                           test_df_full.query("ts_index in [1,2]"),
                                                           lag_n)

unseen_test_preds_catch22_clust4 = compute_lgbm_test_preds(catch22_mod4, 
                                                           test_df_full.query("ts_index in [3,4]"),
                                                           lag_n)

# Compute the forecast performance for each cluster/model, 
# subsetting the test_df to the proper indexes in the function call
catch22_test_perf_clust1 = compute_lgbm_test_perf(unseen_test_preds_catch22_clust1, 
                                                  test_df_full.query("ts_index in [1,2]"))

catch22_test_perf_clust4 = compute_lgbm_test_perf(unseen_test_preds_catch22_clust4, 
                                                  test_df_full.query("ts_index in [3,4]"))

# Append the test set performance data frames into one data frame
catch22_test_perf = catch22_test_perf_clust1.append(catch22_test_perf_clust4)

# Compute the scaled point forecast metrics
catch22_test_perf['nrmse'] = catch22_test_perf['rmse']/catch22_test_perf['mean']
catch22_test_perf['smae'] = catch22_test_perf['mae']/catch22_test_perf['mean']

# Print the point forecast metrics
catch22_test_perf.mean()

# Compute bootstrap PIs for each cluster's model and predictions
catch22_boot_ints_clust1 = compute_lgbm_boostrap_int(unseen_test_preds_catch22_clust1,
                                                     catch22_res1,
                                                     nboot)

catch22_boot_ints_clust4 = compute_lgbm_boostrap_int(unseen_test_preds_catch22_clust4,
                                                     catch22_res4,
                                                     nboot)

# Append the PI data frames for each cluster into one data frame
catch22_boot_ints = catch22_boot_ints_clust1.append(catch22_boot_ints_clust4)

# Add the true values as their own column to the PI data frame
catch22_boot_ints['actual'] = test_df_full.iloc[:,0].to_list()

# Compute the interval scores
catch22_boot_ints['int_95_score'] = interval_score(catch22_boot_ints.actual, 
                                                   catch22_boot_ints.lo_95,
                                                   catch22_boot_ints.hi_95,
                                                   0.95)
                                                    
catch22_boot_ints['int_80_score'] = interval_score(catch22_boot_ints.actual, 
                                                   catch22_boot_ints.lo_80,
                                                   catch22_boot_ints.hi_80,
                                                   0.80)

# Print the average interval scores
catch22_boot_ints.mean()

# Group the PI data by ts_index and then compute the scaled interval scores
catch22_boot_ints_group = catch22_boot_ints.groupby("ts_index").mean().reset_index()
catch22_boot_ints_group['mean'] = catch22_test_perf['mean'].values
catch22_boot_ints_group['int_95_score_scaled'] = catch22_boot_ints_group['int_95_score']/catch22_boot_ints_group['mean']
catch22_boot_ints_group['int_80_score_scaled'] = catch22_boot_ints_group['int_80_score']/catch22_boot_ints_group['mean']

# Print the mean scaled interval scores
catch22_boot_ints_group[['int_95_score_scaled', 'int_80_score_scaled']].mean()

# TSFeat KMeans Clusters

# Check the cluster assignments
unseen_sen_clust.tsfeat

# Load the model for cluster 1 based on the cluster assignments
tsfeat_mod1 = joblib.load("Results/Global/LightGBM Default/TSFeat KMeans/model_1")

# Read in the residuals, convert the data type of the residual lists in the data frame
tsfeat_res = pd.read_csv("Results/Global/LightGBM Default/TSFeat KMeans/residual.csv")
tsfeat_res['residual'] = tsfeat_res['residual'].apply(eval)

# Extract the residuals for the cluster 1 model and save in a list
tsfeat_res1 = tsfeat_res.query("cluster==1")['residual'].values[0]

# Compute test preds with this model
unseen_test_preds_tsfeat1 = compute_lgbm_test_preds(tsfeat_mod1, 
                                                    test_df_full,
                                                    lag_n)

# Compute test pred performance
tsfeat_test_perf = compute_lgbm_test_perf(unseen_test_preds_tsfeat1, test_df_full)

# Add scaled metrics to the test pred performance df
tsfeat_test_perf['nrmse'] = tsfeat_test_perf['rmse']/tsfeat_test_perf['mean']
tsfeat_test_perf['smae'] = tsfeat_test_perf['mae']/tsfeat_test_perf['mean']

# Print the per time series means for point forecast metrics
tsfeat_test_perf.mean()

# Compute bootstrap PIs for this model
tsfeat_boot_ints = compute_lgbm_boostrap_int(unseen_test_preds_tsfeat1,
                                             tsfeat_res1,
                                             nboot)

# Add the true vales to the PI data frame
tsfeat_boot_ints['actual'] = test_df_full.iloc[:,0].to_list()

# Compute the PI scores for each forecast
tsfeat_boot_ints['int_95_score'] = interval_score(tsfeat_boot_ints.actual, 
                                                  tsfeat_boot_ints.lo_95,
                                                  tsfeat_boot_ints.hi_95,
                                                  0.95)
                                                    
tsfeat_boot_ints['int_80_score'] = interval_score(tsfeat_boot_ints.actual, 
                                                  tsfeat_boot_ints.lo_80,
                                                  tsfeat_boot_ints.hi_80,
                                                  0.80)

# Compute the PI score means
tsfeat_boot_ints.mean()

# Group by ts_index and compute scaled PI scores for each time series

# First, compute the mean interval score for each time series
tsfeat_boot_ints_group = tsfeat_boot_ints.groupby("ts_index").mean().reset_index()
# Add the true mean to the data frame
tsfeat_boot_ints_group['mean'] = tsfeat_test_perf['mean'].values
# Compute the scaled interval score for each time series
tsfeat_boot_ints_group['int_95_score_scaled'] = tsfeat_boot_ints_group['int_95_score']/tsfeat_boot_ints_group['mean']
tsfeat_boot_ints_group['int_80_score_scaled'] = tsfeat_boot_ints_group['int_80_score']/tsfeat_boot_ints_group['mean']

# Print mean scaled PI scores
tsfeat_boot_ints_group[['int_95_score_scaled', 'int_80_score_scaled']].mean()

# DTW Clusters

# Check the cluster assignments
unseen_sen_clust.dtw

# Load the models needed based on the cluster assignments
dtw_mod2 = joblib.load("Results/Global/LightGBM Default/DTW/model_2")
dtw_mod1 = joblib.load("Results/Global/LightGBM Default/DTW/model_1")

# Load the residual data frame and cast each model's residual list to the correct data type
dtw_res = pd.read_csv("Results/Global/LightGBM Default/DTW/residual.csv")
dtw_res['residual'] = dtw_res['residual'].apply(eval)

# Extract the residuals for each cluster's model and save in a list
dtw_res2 = dtw_res.query("cluster==2")['residual'].values[0]
dtw_res1 = dtw_res.query("cluster==1")['residual'].values[0]

# Compute test preds for each cluster/model, subset the test df to the proper ts_index in the function call
unseen_test_preds_dtw2 = compute_lgbm_test_preds(dtw_mod2, 
                                                 test_df_full.query("ts_index in [1,2]"),
                                                 lag_n)

unseen_test_preds_dtw1 = compute_lgbm_test_preds(dtw_mod1, 
                                                 test_df_full.query("ts_index in [3,4]"),
                                                 lag_n)

# Compute test pred performance for each cluster
dtw_test_perf2 = compute_lgbm_test_perf(unseen_test_preds_dtw2, test_df_full.query("ts_index in [1,2]"))
dtw_test_perf1 = compute_lgbm_test_perf(unseen_test_preds_dtw1, test_df_full.query("ts_index in [3,4]"))

# Append the data frames for each cluster together
dtw_test_perf = dtw_test_perf2.append(dtw_test_perf1)

# Compute scaled forecast metrics
dtw_test_perf['nrmse'] = dtw_test_perf['rmse']/dtw_test_perf['mean']
dtw_test_perf['smae'] = dtw_test_perf['mae']/dtw_test_perf['mean']

# Print point forecast metrics
dtw_test_perf.mean()

# Compute the bootstrap PIs for each cluster/model
dtw_boot_ints2 = compute_lgbm_boostrap_int(unseen_test_preds_dtw2,
                                           dtw_res2,
                                           nboot)

dtw_boot_ints1 = compute_lgbm_boostrap_int(unseen_test_preds_dtw1,
                                           dtw_res1,
                                           nboot)

# Append the PI data frames into one and add the true values as a column to the data frame
dtw_boot_ints = dtw_boot_ints2.append(dtw_boot_ints1)
dtw_boot_ints['actual'] = test_df_full.iloc[:,0].to_list()

# Compute PI score for each forecast
dtw_boot_ints['int_95_score'] = interval_score(dtw_boot_ints.actual, 
                                               dtw_boot_ints.lo_95,
                                               dtw_boot_ints.hi_95,
                                               0.95)
                                                    
dtw_boot_ints['int_80_score'] = interval_score(dtw_boot_ints.actual, 
                                               dtw_boot_ints.lo_80,
                                               dtw_boot_ints.hi_80,
                                               0.80)

# Print PI score means
dtw_boot_ints.mean()

# Compute scaled PI score for each individual time series in the data set
dtw_boot_ints_group = dtw_boot_ints.groupby("ts_index").mean().reset_index()
dtw_boot_ints_group['mean'] = dtw_test_perf['mean'].values
dtw_boot_ints_group['int_95_score_scaled'] = dtw_boot_ints_group['int_95_score']/dtw_boot_ints_group['mean']
dtw_boot_ints_group['int_80_score_scaled'] = dtw_boot_ints_group['int_80_score']/dtw_boot_ints_group['mean']

# Compute per time series mean of the scaled interval scores
dtw_boot_ints_group[['int_95_score_scaled', 'int_80_score_scaled']].mean()
