==> 17. Unseen Sensor Forecast with DeepAR.ipynb <==
# Both mxnet and gluonts should be installed, in that order, using pip, not conda. The commandas are
# %pip install mxnet
# %pip install gluonts

# To install specific versions, run 
# %pip install mxnet==1.9.1
# %pip install gluonts==0.9.4

# For other libraries, the below can be followed:
# If the libraries are not yet installed, they can be installed in this notebook using commands similar to the below
# %conda install numpy
# %conda install pandas
# %conda install matplotlib
# %conda install scikit-learn
# %conda install joblib
# %conda install tdqm

# Something like the following may also work if the above does not
# import sys
# !conda install --yes --prefix {sys.prefix} numpy
# !conda install --yes --prefix {sys.prefix} pandas
# !conda install --yes --prefix {sys.prefix} scikit-learn
# !conda install --yes --prefix {sys.prefix} joblib
# !conda install --yes --prefix {sys.prefix} tdqm

# To install a specific version, add the version to the install command
# E.g., %conda install numpy=1.20.3

# If all else fails, use pip or follow additional advice such as found at
# https://jakevdp.github.io/blog/2017/12/05/installing-python-packages-from-jupyter/

# If your plan to use pip (especially if you are not working within a specified conda environment), 
# the pip commands might look like:
# pip install numpy
# pip install pandas
# pip install scikit-learn
# pip install joblib
# pip install tdqm

# To install a specific version, add the version to the pip install command
# E.g., pip install numpy==1.20.3

# Load all necessary libraries 
import time
import numpy as np
import pandas as pd
import json
import matplotlib.pyplot as plt
import itertools
from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error
import glob
import random

import mxnet as mx
from gluonts.model import deepar
from gluonts.mx.trainer import Trainer
from gluonts.dataset.common import ListDataset
from gluonts.evaluation import make_evaluation_predictions

from joblib import Parallel, delayed
import contextlib
import joblib
from tqdm import tqdm
import pickle
import os

# Set seeds for reproducibility 
# Make sure we set the mxnet seed for DeepAR models in addition to the other seeds used previously
random.seed(54321)
np.random.seed(54321)
mx.random.seed(54321)

# Read in Data and Prepare for Modeling

# Read in all csv files for England from the Processed directory and save in a list

# Create an empty list
total_df_list = list()

# Loop through all file names in alphebtical order to match R list.files function, 
# read into a pandas df, ensure each df is in chronological order, and append to list
for fname in sorted(glob.glob("Data/Unseen Sensor/Processed/*.csv")):
    print("Reading {}".format(fname))
    df = pd.read_csv(fname)
    df = df.sort_values(by="timestamp")
    total_df_list.append(df)

# Read in the start and end points csv, and subtract 1 to deal with index differences between R and python
start_end = pd.read_csv("start_end_points_unseen.csv")
start_end["start"] = start_end["start"] - 1
start_end["end"] = start_end["end"]

# Initialize an empty list
subset_df_list = list()

# For each df in the total_df_list
for idx, df in enumerate(total_df_list):
    
    # subset the df using the start and end points
    subset_df = df.iloc[start_end.iloc[idx,0]:start_end.iloc[idx,1], ]\
    .reset_index(drop=True).reset_index(drop=False)\
    .rename(columns={"index":"rn"}) # Create a row_num col based on the index of the data frame
    
    # Create a train_val_test field to denote which set each observation is a part of
    subset_df["train_val_test"] = np.where(subset_df["rn"]<(96*7*8),
                                           "train",
                                           np.where(subset_df["rn"]<(96*7*10),
                                                    "val",
                                                    "test"
                                                   )
                                       )

    # Append the subset df to the list
    subset_df_list.append(subset_df)

# Create two lists, one for the train_val data and one for the test data
train_df_list = list()
test_df_list = list()

# Loop through the list of subset dfs
for df in subset_df_list:
    
    # Filter to train and val only for the training data
    train_df = df.query("train_val_test != 'test'").copy()
    train_df = train_df[['timestamp', 'total_volume']]\
    .rename(columns={'timestamp':'ds', 'total_volume':'y'}) # For prophet, we must rename our timestamp column to ds and out target to y
    
    # Append to list
    train_df_list.append(train_df)
    
    # Same steps for test df list
    test_df = df.query("train_val_test == 'test'").copy()
    test_df = test_df[['timestamp', 'total_volume']]\
    .rename(columns={'timestamp':'ds', 'total_volume':'y'})
    test_df_list.append(test_df)

# Set the prediction and context length to match what was used previously
prediction_length = 1
context_length = 672

# Read in the cluster assignments for the new sensor data
clust_assign = pd.read_csv("Results/Unseen Sensor/clust_assign.csv")

# Define Functions for Making Forecasts

# Code for progress bar:
# https://stackoverflow.com/questions/24983493/tracking-progress-of-joblib-parallel-execution
# This allows us to print a progress bar while running parallel loops using joblib 

@contextlib.contextmanager
def tqdm_joblib(tqdm_object):
    """Context manager to patch joblib to report into tqdm progress bar given as argument"""
    class TqdmBatchCompletionCallback(joblib.parallel.BatchCompletionCallBack):
        def __call__(self, *args, **kwargs):
            tqdm_object.update(n=self.batch_size)
            return super().__call__(*args, **kwargs)

    old_batch_callback = joblib.parallel.BatchCompletionCallBack
    joblib.parallel.BatchCompletionCallBack = TqdmBatchCompletionCallback
    try:
        yield tqdm_object
    finally:
        joblib.parallel.BatchCompletionCallBack = old_batch_callback
        tqdm_object.close()

# Function to compute clustered test preds
def compute_deepar_test_preds_clust(model, test_df_ls, ts_idx_ls):
    """Function which takes inputs: a trained model, a list of test data frames, and a time series index list,
    and which returns a tuple of model predictions on the test data"""
        
    # Set up empty dictionaries to save results in
    forecast_means = dict()
    forecast_medians = dict()
    forecast_q10s = dict()
    forecast_q90s = dict()
    forecast_q025s = dict()
    forecast_q975s = dict()
    
    # Rename the loaded model to mod
    mod = model
    
    # For each time series index in the provided list, add the index as a key to the empty dictionaries
    # and add a en empty list as the value
    for idx in ts_idx_ls:
        forecast_means[idx] = list()
        forecast_medians[idx] = list()
        forecast_q10s[idx] = list()
        forecast_q90s[idx] = list()
        forecast_q025s[idx] = list()
        forecast_q975s[idx] = list()

    # Loop through the 2 weeks of data in the test set (we could change this to not be harcoded if the time frame
    # for prediction changed in the future)
    for n in range(1344):
        # Create a list of dictionaries of the start time and total set of target values up to the nth value in the test set
        # Each dict in the list corresponds to one time series
        eval_dict_ls = [{"start": df.iloc[0,5],
                         "target": df.iloc[:,[5,8]].iloc[0:(96*10*7+n),:].total_volume.values
                        } for df in test_df_ls]
        # Create a GluonTS data set, specifiying the frequency of the data as 15 minutes
        eval_data = ListDataset(eval_dict_ls, freq="15min")
        
        # Set the random seeds to make predictions more reproducible
        np.random.seed(n)
        mx.random.seed(n)
        
        # Compute the model predictions, using 100 sample paths for evaluating intervals
        forecast_it, ts_it = make_evaluation_predictions(dataset=eval_data,  # test dataset
                                                         predictor=mod,  # predictor
                                                         num_samples=100,  # number of sample paths we want for evaluation
                                                        )

        # Create a list of forecasts from the forecast_it object
        forecasts = list(forecast_it)

        # Loop through the forecasts
        for i in range(len(forecasts)):
            # Compute the forecast mean
            forecast_mean = forecasts[i].samples.mean()
            # Add the mean to the list in the dictionary of means with a key correspoding to the time series index
            forecast_means[ts_idx_ls[i]].append(forecast_mean)

            # Compute the percentiles of each forecast to use for prediction intervals
            forecast_percentiles = np.percentile(forecasts[i].samples, [2.5, 10, 50, 90, 97.5])
            forecast_medians[ts_idx_ls[i]].append(forecast_percentiles[2])

            # Append the percentile values to the lists within each dictionary for each time series index
            forecast_q10s[ts_idx_ls[i]].append(forecast_percentiles[1])
            forecast_q90s[ts_idx_ls[i]].append(forecast_percentiles[3])
            forecast_q025s[ts_idx_ls[i]].append(forecast_percentiles[0])
            forecast_q975s[ts_idx_ls[i]].append(forecast_percentiles[4])       
           
    # Return a tuple of the dictionaries containing the forecast means and percentiles
    return (forecast_means, forecast_medians, forecast_q10s, forecast_q90s, forecast_q025s, forecast_q975s)

# Create a function to compute the interval score
def interval_score(true_values, lower, upper, interval_range):
    """ Function which takes in the true values, the upper and lower bounds of PIs, and the PI level (e.g., 90%)
        and from these inputs, computes the interval score for each prediction
    """
    
    # Compute alpha from the interval range
    alpha = 100*(1-interval_range)
    
    # Save the upper, lower, and true_values as numpy arrays for computation purposes
    upper = np.array(upper)
    lower = np.array(lower)
    true_values = np.array(true_values)
    
    # Compute the lower component of the interval score - just a boolean for true below interval
    def lower_ind(true,low):
        if true<low:
            return 1
        else:
            return 0
        
    # Computer the upper component of the interval score - similar boolean for true above interval
    def upper_ind(true,up):
        if true>up:
            return 1
        else:
            return 0
        
    # Computer the actual score for each obsveration - formula here: https://epiforecasts.io/scoringutils/reference/interval_score.html
    scores = (upper-lower) + (2/alpha)*(lower-true_values)*(lower > true_values) + (2/alpha)*(true_values-upper)*(true_values > upper)
    
    # Return the scores array
    return scores

# Forecast with Global Model for All Data

# Load the full model via pickle
with open('Results/Global/DeepAR/Full/mod', 'rb') as mod_file:
    mod = pickle.load(mod_file)

# Compute the predictions for each time series using the DeepAR full model. This is done in parallel, where
# the "list" of test data frames passed to the models is really just a list with one data frame in it

# These predictions are returned as list entries, so each entry in the list is a tuple
with tqdm_joblib(tqdm(desc="Full Model DeepAR Unseen Preds", 
                      total=(len(subset_df_list)))) as progress_bar:
    full_mod_preds = Parallel(n_jobs=4)(delayed(compute_deepar_test_preds_clust)(mod,
                                                                                 [subset_df_list[i]],
                                                                                 [i+1]
                                                                                ) for i in range(len(subset_df_list)))

# We create a new dictionary for each of the forecast means and percentiles by adding the dictionary
# from each of the tuples returned in the last cell
forec_means = {**full_mod_preds[0][0], 
               **full_mod_preds[1][0],
               **full_mod_preds[2][0], 
               **full_mod_preds[3][0]}

forec_lo_80 = {**full_mod_preds[0][2], 
               **full_mod_preds[1][2],
               **full_mod_preds[2][2], 
               **full_mod_preds[3][2]}

forec_hi_80 = {**full_mod_preds[0][3], 
               **full_mod_preds[1][3],
               **full_mod_preds[2][3], 
               **full_mod_preds[3][3]}

forec_lo_95 = {**full_mod_preds[0][4], 
               **full_mod_preds[1][4],
               **full_mod_preds[2][4], 
               **full_mod_preds[3][4]}

forec_hi_95 = {**full_mod_preds[0][5], 
               **full_mod_preds[1][5],
               **full_mod_preds[2][5], 
               **full_mod_preds[3][5]}


# We create a dictionary of the true values, 
# where the keys are the ts_index and the values are the lists of true values
forec_actuals = {k: test_df_list[k-1].y.to_list() for k in range(1, (len(test_df_list)+1))}

# Create a dictionary of acutal means, where the keys are the ts_indexes and the values are the true test set means
forec_actual_means = {k: np.mean(test_df_list[k-1].y.to_list()) for k in range(1, (len(test_df_list)+1))}

# Create dictionaries are the keys are the ts_indexes and the values are the lists of interval scores for each forecast
# We do this for the 80% and 95% PIs
full_int_score_80 = {k: interval_score(forec_actuals[k], 
                                       forec_lo_80[k],
                                       forec_hi_80[k],
                                       0.8) for k in forec_actuals.keys()
                    }

full_int_score_95 = {k: interval_score(forec_actuals[k], 
                                       forec_lo_95[k],
                                       forec_hi_95[k],
                                       0.95) for k in forec_actuals.keys()
                    }

# Create a dictionary of RMSE scores, where the keys are the indexes and the values are the RMSE
full_rmse = {k: mean_squared_error(forec_actuals[k], 
                                   forec_means[k], 
                                   squared=False) for k in forec_actuals.keys()}

# Create a dictionary of MAE scores, where the keys are the indexes and the values are the MAE
full_mae = {k: mean_absolute_error(forec_actuals[k], 
                                   forec_means[k]) for k in forec_actuals.keys()}

# Compute and print the per time series means of the RMSE
np.mean(list(full_rmse.values()))

# Compute and print the per time series means of the MAE
np.mean(list(full_mae.values()))

# Compute and print the scaled RMSE mean
np.mean([full_rmse[k]/forec_actual_means[k] for k in full_rmse.keys()])

# Compute and print the scaled MAE mean
np.mean([full_mae[k]/forec_actual_means[k] for k in full_mae.keys()])

# Compute and print the mean interval scores
np.mean(list(full_int_score_80.values()))

np.mean(list(full_int_score_95.values()))

# Compute and print the scaled interval scores
np.mean([full_int_score_80[k]/forec_actual_means[k] for k in full_int_score_80.keys()])

np.mean([full_int_score_95[k]/forec_actual_means[k] for k in full_int_score_95.keys()])

# Forecast with Models for Random Clusters

# Read in cluster assignments
clust_assign.rand

# Load models for each cluster
with open(f'Results/Global/DeepAR/Random Cluster/mod_2', 'rb') as f:
    rand_clust_deep_ar_mod3 = pickle.load(f)
    
with open(f'Results/Global/DeepAR/Random Cluster/mod_1', 'rb') as f:
    rand_clust_deep_ar_mod2 = pickle.load(f)
    
with open(f'Results/Global/DeepAR/Random Cluster/mod_0', 'rb') as f:
    rand_clust_deep_ar_mod1 = pickle.load(f)
    
with open(f'Results/Global/DeepAR/Random Cluster/mod_3', 'rb') as f:
    rand_clust_deep_ar_mod4 = pickle.load(f)

# Save the models in a list, in the same order as the time series for which they will be used
rand_clust_deep_ar_mod_ls = [rand_clust_deep_ar_mod3, rand_clust_deep_ar_mod2, 
                             rand_clust_deep_ar_mod1, rand_clust_deep_ar_mod4]

# Loop through the list of models and the test set for each time series and compute predictions, in parallel
with tqdm_joblib(tqdm(desc="Random Cluster DeepAR Unseen Preds", 
                      total=(len(subset_df_list)))) as progress_bar:
    rand_clust_test_preds = Parallel(n_jobs=4)(delayed(compute_deepar_test_preds_clust)(rand_clust_deep_ar_mod_ls[i],
                                                                                        [subset_df_list[i]],
                                                                                        [i+1]
                                                                                       ) for i in range(len(subset_df_list)))

# As before, create dictionaries from the list of tuples (which contain dictionaries) created in the last cell
rand_clust_forec_means = {**rand_clust_test_preds[0][0], 
                          **rand_clust_test_preds[1][0],
                          **rand_clust_test_preds[2][0], 
                          **rand_clust_test_preds[3][0]}

rand_clust_forec_lo_80 = {**rand_clust_test_preds[0][2], 
                          **rand_clust_test_preds[1][2],
                          **rand_clust_test_preds[2][2], 
                          **rand_clust_test_preds[3][2]}

rand_clust_forec_hi_80 = {**rand_clust_test_preds[0][3], 
                          **rand_clust_test_preds[1][3],
                          **rand_clust_test_preds[2][3], 
                          **rand_clust_test_preds[3][3]}

rand_clust_forec_lo_95 = {**rand_clust_test_preds[0][4], 
                          **rand_clust_test_preds[1][4],
                          **rand_clust_test_preds[2][4], 
                          **rand_clust_test_preds[3][4]}

rand_clust_forec_hi_95 = {**rand_clust_test_preds[0][5], 
                          **rand_clust_test_preds[1][5],
                          **rand_clust_test_preds[2][5], 
                          **rand_clust_test_preds[3][5]}

rand_clust_forec_actuals = {k: test_df_list[k-1].y.to_list() for k in range(1, (len(test_df_list)+1))}

rand_clust_forec_actual_means = {k: np.mean(test_df_list[k-1].y.to_list()) for k in range(1, (len(test_df_list)+1))}

# Again, create dictionaries of interval scores where each entry is for one time series in the set
rand_clust_int_score_80 = {k: interval_score(rand_clust_forec_actuals[k], 
                                             rand_clust_forec_lo_80[k],
                                             rand_clust_forec_hi_80[k],
                                             0.8) for k in rand_clust_forec_actuals.keys()
                          }

rand_clust_int_score_95 = {k: interval_score(rand_clust_forec_actuals[k], 
                                             rand_clust_forec_lo_95[k],
                                             rand_clust_forec_hi_95[k],
                                             0.95) for k in rand_clust_forec_actuals.keys()
                          }

# Again, create dictionaries of point forecast performance metrics
rand_clust_rmse = {k: mean_squared_error(rand_clust_forec_actuals[k], 
                                         rand_clust_forec_means[k], 
                                         squared=False) for k in rand_clust_forec_actuals.keys()}

rand_clust_mae = {k: mean_absolute_error(rand_clust_forec_actuals[k], 
                                         rand_clust_forec_means[k]) for k in rand_clust_forec_actuals.keys()}

# Print raw and scaled point forecast metrics
np.mean(list(rand_clust_rmse.values()))

np.mean(list(rand_clust_mae.values()))

np.mean([rand_clust_rmse[k]/rand_clust_forec_actual_means[k] for k in rand_clust_rmse.keys()])

np.mean([rand_clust_mae[k]/rand_clust_forec_actual_means[k] for k in rand_clust_mae.keys()])

# Print raw and scaled interval scores
np.mean(list(rand_clust_int_score_80.values()))

np.mean(list(rand_clust_int_score_95.values()))

np.mean([rand_clust_int_score_80[k]/rand_clust_forec_actual_means[k] for k in rand_clust_int_score_80.keys()])

np.mean([rand_clust_int_score_95[k]/rand_clust_forec_actual_means[k] for k in rand_clust_int_score_95.keys()])

# Forecast with Highway System Model

# Load in the model for the highways england data
with open(f'Results/Global/DeepAR/Highway System/mod_0', 'rb') as f:
    eng_mod = pickle.load(f)

# In parallel, loop through the time series in the test set and compute predictions
with tqdm_joblib(tqdm(desc="England Model DeepAR Unseen Preds", 
                      total=(len(subset_df_list)))) as progress_bar:
    eng_test_preds_total_ls = Parallel(n_jobs=4)(delayed(compute_deepar_test_preds_clust)(eng_mod,
                                                                                          [subset_df_list[i]],
                                                                                          [i+1]
                                                                                         ) for i in range(len(subset_df_list))) 


# Again, create dictionaries of the model forecasts (should probably create a function for this instead of doing it
# by hand each time)
eng_forec_means = {**eng_test_preds_total_ls[0][0], 
                   **eng_test_preds_total_ls[1][0],
                   **eng_test_preds_total_ls[2][0], 
                   **eng_test_preds_total_ls[3][0]}

eng_forec_lo_80 = {**eng_test_preds_total_ls[0][2], 
                   **eng_test_preds_total_ls[1][2],
                   **eng_test_preds_total_ls[2][2], 
                   **eng_test_preds_total_ls[3][2]}

eng_forec_hi_80 = {**eng_test_preds_total_ls[0][3], 
                   **eng_test_preds_total_ls[1][3],
                   **eng_test_preds_total_ls[2][3], 
                   **eng_test_preds_total_ls[3][3]}

eng_forec_lo_95 = {**eng_test_preds_total_ls[0][4], 
                   **eng_test_preds_total_ls[1][4],
                   **eng_test_preds_total_ls[2][4], 
                   **eng_test_preds_total_ls[3][4]}

eng_forec_hi_95 = {**eng_test_preds_total_ls[0][5], 
                   **eng_test_preds_total_ls[1][5],
                   **eng_test_preds_total_ls[2][5], 
                   **eng_test_preds_total_ls[3][5]}

eng_forec_actuals = {k: test_df_list[k-1].y.to_list() for k in range(1, (len(test_df_list)+1))}

eng_forec_actual_means = {k: np.mean(test_df_list[k-1].y.to_list()) for k in range(1, (len(test_df_list)+1))}

# Compute the PI scores for each forecast
eng_int_score_80 = {k: interval_score(eng_forec_actuals[k], 
                                      eng_forec_lo_80[k],
                                      eng_forec_hi_80[k],
                                      0.8) for k in eng_forec_actuals.keys()
                   }

eng_int_score_95 = {k: interval_score(eng_forec_actuals[k], 
                                      eng_forec_lo_95[k],
                                      eng_forec_hi_95[k],
                                      0.95) for k in eng_forec_actuals.keys()
                   }

# Compute point forecast metrics
eng_rmse = {k: mean_squared_error(eng_forec_actuals[k], 
                                  eng_forec_means[k], 
                                  squared=False) for k in eng_forec_actuals.keys()}

eng_mae = {k: mean_absolute_error(eng_forec_actuals[k], 
                                  eng_forec_means[k]) for k in eng_forec_actuals.keys()}

# Print the point forecast metrics
np.mean(list(eng_rmse.values()))

np.mean(list(eng_mae.values()))

np.mean([eng_rmse[k]/eng_forec_actual_means[k] for k in eng_rmse.keys()])

np.mean([eng_mae[k]/eng_forec_actual_means[k] for k in eng_mae.keys()])

# Print the interval scores
np.mean(list(eng_int_score_80.values()))

np.mean(list(eng_int_score_95.values()))

np.mean([eng_int_score_80[k]/eng_forec_actual_means[k] for k in eng_int_score_80.keys()])

np.mean([eng_int_score_95[k]/eng_forec_actual_means[k] for k in eng_int_score_95.keys()])

# Forecast with Models for Catch22-Based KMeans Clusters

# Print the cluster assignments for the next set of clusters
clust_assign.catch22

# Load in the relevant model objects
with open(f'Results/Global/DeepAR/Catch22 KMeans/mod_0', 'rb') as f:
    catch22_mod1 = pickle.load(f)
    
with open(f'Results/Global/DeepAR/Catch22 KMeans/mod_3', 'rb') as f:
    catch22_mod4 = pickle.load(f)

# Save the models to a list, repeating models where necessary so there is one entry in the list per time series
# in the data set where the order of the models in the list is the same as the time series index based 
# order of cluster assignments
catch22_deepar_mod_ls = [catch22_mod1, catch22_mod1,
                         catch22_mod4, catch22_mod4] 

# Loop through the list of models and the test set dfs and compute predictions, in parallel
with tqdm_joblib(tqdm(desc="Catch22 DeepAR Unseen Preds", 
                      total=(len(subset_df_list)))) as progress_bar:
    catch22_test_preds = Parallel(n_jobs=4)(delayed(compute_deepar_test_preds_clust)(catch22_deepar_mod_ls[i],
                                                                                     [subset_df_list[i]],
                                                                                     [i+1]
                                                                                    ) for i in range(len(subset_df_list))) 


# Again, create dictionaries of forecast means, percentiles, and true values where each entry is for one time series
catch22_forec_means = {**catch22_test_preds[0][0], 
                       **catch22_test_preds[1][0],
                       **catch22_test_preds[2][0], 
                       **catch22_test_preds[3][0]}

catch22_forec_lo_80 = {**catch22_test_preds[0][2], 
                       **catch22_test_preds[1][2],
                       **catch22_test_preds[2][2], 
                       **catch22_test_preds[3][2]}

catch22_forec_hi_80 = {**catch22_test_preds[0][3], 
                       **catch22_test_preds[1][3],
                       **catch22_test_preds[2][3], 
                       **catch22_test_preds[3][3]}

catch22_forec_lo_95 = {**catch22_test_preds[0][4], 
                       **catch22_test_preds[1][4],
                       **catch22_test_preds[2][4], 
                       **catch22_test_preds[3][4]}

catch22_forec_hi_95 = {**catch22_test_preds[0][5], 
                       **catch22_test_preds[1][5],
                       **catch22_test_preds[2][5], 
                       **catch22_test_preds[3][5]}

catch22_forec_actuals = {k: test_df_list[k-1].y.to_list() for k in range(1, (len(test_df_list)+1))}

catch22forec_actual_means = {k: np.mean(test_df_list[k-1].y.to_list()) for k in range(1, (len(test_df_list)+1))}

# Compute the interval scores for each forecast
catch22_int_score_80 = {k: interval_score(catch22_forec_actuals[k], 
                                          catch22_forec_lo_80[k],
                                          catch22_forec_hi_80[k],
                                          0.8) for k in catch22_forec_actuals.keys()
                       }

catch22_int_score_95 = {k: interval_score(catch22_forec_actuals[k], 
                                          catch22_forec_lo_95[k],
                                          catch22_forec_hi_95[k],
                                          0.95) for k in catch22_forec_actuals.keys()
                       }

# Compute the RMSE and MAE values for the forecasts
catch22_rmse = {k: mean_squared_error(catch22_forec_actuals[k], 
                                      catch22_forec_means[k], 
                                      squared=False) for k in catch22_forec_actuals.keys()}

catch22_mae = {k: mean_absolute_error(catch22_forec_actuals[k], 
                                      catch22_forec_means[k]) for k in catch22_forec_actuals.keys()}

# Print the point forecast scores
np.mean(list(catch22_rmse.values()))

np.mean(list(catch22_mae.values()))

np.mean([catch22_rmse[k]/catch22forec_actual_means[k] for k in catch22_rmse.keys()])

np.mean([catch22_mae[k]/catch22forec_actual_means[k] for k in catch22_mae.keys()])

# Print the interval scores
np.mean(list(catch22_int_score_80.values()))

np.mean(list(catch22_int_score_95.values()))

np.mean([catch22_int_score_80[k]/catch22forec_actual_means[k] for k in catch22_int_score_80.keys()])

np.mean([catch22_int_score_95[k]/catch22forec_actual_means[k] for k in catch22_int_score_95.keys()])

# Forecast for Models with TSFeat-Based KMeans Clusters

# Print the cluster assignments - for this technique, they are all the same
clust_assign.tsfeat

# Load the model for cluster 1
with open(f'Results/Global/DeepAR/TSFeat KMeans/mod_0', 'rb') as f:
    tsfeat_mod1 = pickle.load(f)

# Create a list where all entries are model 1 (this is probably not needed)
ts_feat_mod_ls = [tsfeat_mod1, tsfeat_mod1, tsfeat_mod1, tsfeat_mod1]

# Compute the model predictions and save the tuple of predictions for each time series as a list entry
with tqdm_joblib(tqdm(desc="TSFEat DeepAR Unseen Preds", 
                      total=(len(subset_df_list)))) as progress_bar:
    tsfeat_test_preds = Parallel(n_jobs=4)(delayed(compute_deepar_test_preds_clust)(ts_feat_mod_ls[i],
                                                                                    [subset_df_list[i]],
                                                                                    [i+1]
                                                                                   ) for i in range(len(subset_df_list))) 


# Create dictionaries from the above tuples of dictionaries for the forecast means, PIs, and true means 
tsfeat_forec_means = {**tsfeat_test_preds[0][0], 
                      **tsfeat_test_preds[1][0],
                      **tsfeat_test_preds[2][0], 
                      **tsfeat_test_preds[3][0]}

tsfeat_forec_lo_80 = {**tsfeat_test_preds[0][2], 
                      **tsfeat_test_preds[1][2],
                      **tsfeat_test_preds[2][2], 
                      **tsfeat_test_preds[3][2]}

tsfeat_forec_hi_80 = {**tsfeat_test_preds[0][3], 
                      **tsfeat_test_preds[1][3],
                      **tsfeat_test_preds[2][3], 
                      **tsfeat_test_preds[3][3]}

tsfeat_forec_lo_95 = {**tsfeat_test_preds[0][4], 
                      **tsfeat_test_preds[1][4],
                      **tsfeat_test_preds[2][4], 
                      **tsfeat_test_preds[3][4]}

tsfeat_forec_hi_95 = {**tsfeat_test_preds[0][5], 
                      **tsfeat_test_preds[1][5],
                      **tsfeat_test_preds[2][5], 
                      **tsfeat_test_preds[3][5]}

tsfeat_forec_actuals = {k: test_df_list[k-1].y.to_list() for k in range(1, (len(test_df_list)+1))}

tsfeatforec_actual_means = {k: np.mean(test_df_list[k-1].y.to_list()) for k in range(1, (len(test_df_list)+1))}

# Compute the interval scores
tsfeat_int_score_80 = {k: interval_score(tsfeat_forec_actuals[k], 
                                         tsfeat_forec_lo_80[k],
                                         tsfeat_forec_hi_80[k],
                                         0.8) for k in tsfeat_forec_actuals.keys()
                      }

tsfeat_int_score_95 = {k: interval_score(tsfeat_forec_actuals[k], 
                                         tsfeat_forec_lo_95[k],
                                         tsfeat_forec_hi_95[k],
                                         0.95) for k in tsfeat_forec_actuals.keys()
                      }

# Compute the point forecast metrics
tsfeat_rmse = {k: mean_squared_error(tsfeat_forec_actuals[k], 
                                     tsfeat_forec_means[k], 
                                     squared=False) for k in tsfeat_forec_actuals.keys()}

tsfeat_mae = {k: mean_absolute_error(tsfeat_forec_actuals[k], 
                                     tsfeat_forec_means[k]) for k in tsfeat_forec_actuals.keys()}

# Print the point forecast metrics
np.mean(list(tsfeat_rmse.values()))

np.mean(list(tsfeat_mae.values()))

# Print the scaled point forecast metrics
np.mean([tsfeat_rmse[k]/tsfeatforec_actual_means[k] for k in tsfeat_rmse.keys()])

np.mean([tsfeat_mae[k]/tsfeatforec_actual_means[k] for k in tsfeat_mae.keys()])

# Print the interval scores
np.mean(list(tsfeat_int_score_80.values()))

np.mean(list(tsfeat_int_score_95.values()))

# Print the scaled interval scores
np.mean([tsfeat_int_score_80[k]/tsfeatforec_actual_means[k] for k in tsfeat_int_score_80.keys()])

np.mean([tsfeat_int_score_95[k]/tsfeatforec_actual_means[k] for k in tsfeat_int_score_95.keys()])

# Forecast with Models for DTW-Based K-Medoids Clusters

# Repeat the process for the final clustering method, first checking the cluster assignments
clust_assign.dtw

# Load the model for each cluster
with open(f'Results/Global/DeepAR/DTW/mod_1', 'rb') as f:
    dtw_mod2 = pickle.load(f)
    
with open(f'Results/Global/DeepAR/DTW/mod_0', 'rb') as f:
    dtw_mod1 = pickle.load(f)

# Create a list with the models in it, the first two entries for model one (time series numbers 1 and 2)
# and the last two entries for model 1 (time series indexes 3 and 4)
dtw_mod_list = [dtw_mod2, dtw_mod2, dtw_mod1, dtw_mod1]

# In parallel, loop through the list of models and the test data and compute forecasts
with tqdm_joblib(tqdm(desc="DTW DeepAR Unseen Preds", 
                      total=(len(subset_df_list)))) as progress_bar:
    dtw_test_preds = Parallel(n_jobs=4)(delayed(compute_deepar_test_preds_clust)(dtw_mod_list[i],
                                                                                 [subset_df_list[i]],
                                                                                 [i+1]
                                                                                ) for i in range(len(subset_df_list))) 


# Just as before, compute new dictionaries with keys as ts_index and values as lists of forecast means and percentiles
dtw_forec_means = {**dtw_test_preds[0][0], 
                   **dtw_test_preds[1][0],
                   **dtw_test_preds[2][0], 
                   **dtw_test_preds[3][0]}

dtw_forec_lo_80 = {**dtw_test_preds[0][2], 
                   **dtw_test_preds[1][2],
                   **dtw_test_preds[2][2], 
                   **dtw_test_preds[3][2]}

dtw_forec_hi_80 = {**dtw_test_preds[0][3], 
                   **dtw_test_preds[1][3],
                   **dtw_test_preds[2][3], 
                   **dtw_test_preds[3][3]}

dtw_forec_lo_95 = {**dtw_test_preds[0][4], 
                   **dtw_test_preds[1][4],
                   **dtw_test_preds[2][4], 
                   **dtw_test_preds[3][4]}

dtw_forec_hi_95 = {**dtw_test_preds[0][5], 
                   **dtw_test_preds[1][5],
                   **dtw_test_preds[2][5], 
                   **dtw_test_preds[3][5]}

# Also compute dictionaries with true values for each time series and the true mean of each time series
dtw_forec_actuals = {k: test_df_list[k-1].y.to_list() for k in range(1, (len(test_df_list)+1))}

dtwforec_actual_means = {k: np.mean(test_df_list[k-1].y.to_list()) for k in range(1, (len(test_df_list)+1))}

# Compute interval scores
dtw_int_score_80 = {k: interval_score(dtw_forec_actuals[k], 
                                      dtw_forec_lo_80[k],
                                      dtw_forec_hi_80[k],
                                      0.8) for k in dtw_forec_actuals.keys()
                   }

dtw_int_score_95 = {k: interval_score(dtw_forec_actuals[k], 
                                      dtw_forec_lo_95[k],
                                      dtw_forec_hi_95[k],
                                      0.95) for k in dtw_forec_actuals.keys()
                   }

# Compute point forecast metrics
dtw_rmse = {k: mean_squared_error(dtw_forec_actuals[k], 
                                  dtw_forec_means[k], 
                                  squared=False) for k in dtw_forec_actuals.keys()}

dtw_mae = {k: mean_absolute_error(dtw_forec_actuals[k], 
                                  dtw_forec_means[k]) for k in dtw_forec_actuals.keys()}

# Print raw and scaled point forecast metrics
np.mean(list(dtw_rmse.values()))

np.mean(list(dtw_mae.values()))

np.mean([dtw_rmse[k]/dtwforec_actual_means[k] for k in dtw_rmse.keys()])

np.mean([dtw_mae[k]/dtwforec_actual_means[k] for k in dtw_mae.keys()])

# Print raw and scaled interval scores
np.mean(list(dtw_int_score_80.values()))

np.mean(list(dtw_int_score_95.values()))

np.mean([dtw_int_score_80[k]/dtwforec_actual_means[k] for k in dtw_int_score_80.keys()])

np.mean([dtw_int_score_95[k]/dtwforec_actual_means[k] for k in dtw_int_score_95.keys()])
